hir_ty

Module closure_analysis

Structs

CaptureInfo

Part of MinCaptureInformationMap; describes the capture kind (&, &mut, move) for a particular capture as well as identifying the part of the source code that triggered this capture to occur.

CaptureSourceStack

CapturedPlace

A composite describing a Place that is captured by a closure.

ClosureData

Place

A Place represents how a value is located in memory. This does not always correspond to a syntactic place expression. For example, when processing a pattern, a Place can be used to refer to the sub-value currently being inspected.

Projection

Enums

BorrowKind

FakeReadCause

The FakeReadCause describes the type of pattern why a FakeRead statement exists.

PlaceBase

ProjectionKind

UpvarCapture

Information describing the capture of an upvar. This is computed during typeck, specifically by regionck.