hir_ty

Module diagnostics

Type inference-based diagnostics.

Modules

decl_check 🔒

Provides validators for names of declarations.

expr 🔒

Various diagnostics for expressions that are collected together in one pass through the body using inference results: mismatched arg counts, missing fields, etc.

match_check 🔒

Validation of matches.

unsafe_check 🔒

Provides validations for unsafe code. Currently checks if unsafe functions are missing unsafe blocks.

Structs

IncorrectCase

Enums

BodyValidationDiagnostic

CaseType

InsideUnsafeBlock

UnsafetyReason

Functions

incorrect_case

missing_unsafe

record_literal_missing_fields

record_pattern_missing_fields

unsafe_operations

unsafe_operations_for_body