hir_ty::diagnostics

Module expr

Various diagnostics for expressions that are collected together in one pass through the body using inference results: mismatched arg counts, missing fields, etc.

Structs

Body 🔒

The body of an item (function, const etc.).

ExprValidator 🔒

FilterMapNextChecker 🔒

MatchArm 🔒

Enums

BodyValidationDiagnostic

Expr 🔒

Pat 🔒

Close relative to rustc’s hir::PatKind

RecordSpread 🔒

Statement 🔒

VariantId 🔒

Functions

missing_match_arms 🔒

record_literal_missing_fields

record_pattern_missing_fields

types_of_subpatterns_do_match 🔒

Type Aliases

ExprId 🔒

LocalFieldId 🔒

PatId 🔒