hir_ty::display

Struct HirFormatter

pub struct HirFormatter<'a, 'db> {

    pub db: &'db dyn HirDatabase,
    pub interner: DbInterner<'db>,
    fmt: &'a mut dyn HirWrite,
    buf: String,
    curr_size: usize,
    max_size: Option<usize>,
    pub entity_limit: Option<usize>,
    show_container_bounds: bool,
    omit_verbose_types: bool,
    closure_style: ClosureStyle,
    display_lifetimes: DisplayLifetime,
    display_kind: DisplayKind,
    display_target: DisplayTarget,
    bounds_formatting_ctx: BoundsFormattingCtx<'db>,
    trait_bounds_need_parens: bool,

}

Fields
db: &'db dyn HirDatabase

The database handle

interner: DbInterner<'db>fmt: &'a mut dyn HirWrite

The sink to write into

buf: String

A buffer to intercept writes with, this allows us to track the overall size of the formatted output.

curr_size: usize

The current size of the formatted output.

max_size: Option<usize>

Size from which we should truncate the output.

entity_limit: Option<usize>

When rendering something that has a concept of “children” (like fields in a struct), this limits how many should be rendered.

show_container_bounds: bool

When rendering functions, whether to show the constraint from the container

omit_verbose_types: boolclosure_style: ClosureStyledisplay_lifetimes: DisplayLifetimedisplay_kind: DisplayKinddisplay_target: DisplayTargetbounds_formatting_ctx: BoundsFormattingCtx<'db>trait_bounds_need_parens: bool

Whether formatting impl Trait1 + Trait2 or dyn Trait1 + Trait2 needs parentheses around it, for example when formatting &(impl Trait1 + Trait2).

Implementations

impl<'db> HirFormatter<'_, 'db>

pub fn start_location_link(&mut self, location: ModuleDefId)

pub fn start_location_link_generic(&mut self, location: GenericParamId)

pub fn end_location_link(&mut self)

fn format_bounds_with<T, F: FnOnce(&mut Self) -> T>( &mut self, target: AliasTy<'db>, format_bounds: F, ) -> T

fn render_region(&self, lifetime: Region<'db>) -> bool

impl<'db> HirFormatter<'_, 'db>

pub fn krate(&self) -> Crate

pub fn edition(&self) -> Edition

pub fn lang_items(&self) -> &'db LangItems

pub fn write_joined<T: HirDisplay<'db>>( &mut self, iter: impl IntoIterator<Item = T>, sep: &str, ) -> Result

pub fn write_fmt(&mut self, args: Arguments<'_>) -> Result

This allows using the write! macro directly with a HirFormatter.

pub fn write_str(&mut self, s: &str) -> Result

pub fn write_char(&mut self, c: char) -> Result

pub fn should_truncate(&self) -> bool

pub fn omit_verbose_types(&self) -> bool

pub fn show_container_bounds(&self) -> bool