Function hir_ty::could_unify_deeply

pub fn could_unify_deeply(
    db: &dyn HirDatabase,
    env: Arc<TraitEnvironment>,
    tys: &Canonical<(Ty, Ty)>
) -> bool

Check if types unify eagerly making sure there are no unresolved goals.

This means that placeholder types are not considered to unify if there are any bounds set on them. For example Option<T> and Option<U> do not unify as we cannot show that T = U