hir_ty::infer::closure

Module analysis

Inferring borrow kinds for upvars

Whenever there is a closure expression, we need to determine how each upvar is used. We do this by initially assigning each upvar an immutable “borrow kind” (see BorrowKind for details) and then “escalating” the kind as needed. The borrow kind proceeds according to the following lattice:

ty::ImmBorrow -> ty::UniqueImmBorrow -> ty::MutBorrow

So, for example, if we see an assignment x = 5 to an upvar x, we will promote its borrow kind to mutable borrow. If we see an &mut x we’ll do the same. Naturally, this applies not just to the upvar, but to everything owned by x, so the result is the same for something like x.f = 5 and so on (presuming x is not a borrowed pointer to a struct). These adjustments are performed in adjust_for_non_move_closure (you can trace backwards through the code from there).

The fact that we are inferring borrow kinds as we go results in a semi-hacky interaction with the way ExprUseVisitor is computing Places. In particular, it will query the current borrow kind as it goes, and we’ll return the current value, but this may get adjusted later. Therefore, in this module, we generally ignore the borrow kind (and derived mutabilities) that ExprUseVisitor returns within Places, since they may be inaccurate. (Another option would be to use a unification scheme, where instead of returning a concrete borrow kind like ty::ImmBorrow, we return a ty::InferBorrow(upvar_id) or something like that, but this would then mean that all later passes would have to check for these figments and report an error, and it just seems like more mess in the end.)

Modules

expr_use_visitor 🔒

A different sort of visitor for walking fn bodies. Unlike the normal visitor, which just walks the entire body in one shot, the ExprUseVisitor determines how expressions are being used.

Structs

InferBorrowKind 🔒

Enums

BorrowKind

PlaceAncestryRelation 🔒

Describe the relationship between the paths of two places eg:

UpvarArgs 🔒

Functions

adjust_for_move_closure 🔒

Truncate deref of any reference.

adjust_for_non_move_closure 🔒

Adjust closure capture just that if taking ownership of data, only move data from enclosing stack frame.

analyze_coroutine_closure_captures 🔒

apply_capture_kind_on_capture_ty 🔒

Returns a Ty that applies the specified capture kind on the provided capture Ty

child_prefix_matches_parent_projections 🔒

determine_capture_info 🔒

At the end, capture_info_a will contain the selected info.

determine_capture_sources 🔒

determine_place_ancestry_relation 🔒

Determines the Ancestry relationship of Place A relative to Place B

enable_precise_capture 🔒

Precise capture is enabled if user is using Rust Edition 2021 or higher. span is the span of the closure.

restrict_capture_precision 🔒

Truncate projections so that the following rules are obeyed by the captured place:

restrict_precision_for_drop_types 🔒

Rust doesn’t permit moving fields out of a type that implements drop

restrict_precision_for_unsafe 🔒

Truncate place so that an unsafe block isn’t required to capture it.

restrict_repr_packed_field_ref_capture 🔒

Truncate the capture so that the place being borrowed is in accordance with RFC 1240, which states that it’s unsafe to take a reference into a struct marked repr(packed).

should_reborrow_from_env_of_parent_coroutine_closure 🔒

Determines whether a child capture that is derived from a parent capture should be borrowed with the lifetime of the parent coroutine-closure’s env.

truncate_capture_for_optimization 🔒

Reduces the precision of the captured place when the precision doesn’t yield any benefit from borrow checking perspective, allowing us to save us on the size of the capture.

truncate_place_to_len_and_update_capture_kind 🔒

Truncates place to have up to len projections. curr_mode is the current required capture kind for the place. Returns the truncated place and the updated required capture kind.

Type Aliases

InferredCaptureInformation 🔒

Intermediate format to store a captured Place and associated CaptureInfo during capture analysis. Information in this map feeds into the minimum capture analysis pass.