hir_ty::infer

Module op

Inference of binary and unary operators.

Enums

BinOpCategory 🔒

Functions

deref_ty_if_possible 🔒

Dereferences a single level of immutable referencing.

is_builtin_binop 🔒

Returns true if this is a built-in arithmetic operation (e.g., u32 + u32, i16x4 == i16x4) and false if these types would have to be overloaded to be legal. There are two reasons that we distinguish builtin operations from overloaded ones (vs trying to drive everything uniformly through the trait system and intrinsics or something like that):

is_op_by_value 🔒

Returns true if the binary operator takes its arguments by value.