hir_ty::infer

Struct InferenceContext

pub(crate) struct InferenceContext<'body, 'db> {

    pub(crate) db: &'db dyn HirDatabase,
    pub(crate) owner: DefWithBodyId,
    pub(crate) body: &'body Body,
    pub(crate) resolver: Resolver<'db>,
    target_features: OnceCell<(TargetFeatures<'db>, TargetFeatureIsSafeInTarget)>,
    pub(crate) unstable_features: MethodResolutionUnstableFeatures,
    pub(crate) edition: Edition,
    pub(crate) generic_def: GenericDefId,
    pub(crate) table: InferenceTable<'db>,
    pub(crate) lang_items: &'db LangItems,
    traits_in_scope: FxHashSet<TraitId>,
    pub(crate) result: InferenceResult,
    tuple_field_accesses_rev: IndexSet<Tys<'db>, BuildHasherDefault<FxHasher>>,
    return_ty: Ty<'db>,
    return_coercion: Option<CoerceMany<'db, 'db>>,
    resume_yield_tys: Option<(Ty<'db>, Ty<'db>)>,
    diverges: Diverges,
    breakables: Vec<BreakableContext<'db>>,
    types: &'db DefaultAny<'db>,
    inside_assignment: bool,
    deferred_cast_checks: Vec<CastCheck<'db>>,
    current_captures: Vec<CapturedItemWithoutTy>,
    current_capture_span_stack: Vec<MirSpan>,
    current_closure: Option<InternedClosureId>,
    closure_dependencies: FxHashMap<InternedClosureId, Vec<InternedClosureId>>,
    deferred_closures: FxHashMap<InternedClosureId, Vec<(Ty<'db>, Ty<'db>, Vec<Ty<'db>>, ExprId)>>,
    diagnostics: Diagnostics,

}

The inference context contains all information needed during type inference.

Fields
db: &'db dyn HirDatabaseowner: DefWithBodyIdbody: &'body Bodyresolver: Resolver<'db>

Generally you should not resolve things via this resolver. Instead create a TyLoweringContext and resolve the path via its methods. This will ensure proper error reporting.

target_features: OnceCell<(TargetFeatures<'db>, TargetFeatureIsSafeInTarget)>unstable_features: MethodResolutionUnstableFeaturesedition: Editiongeneric_def: GenericDefIdtable: InferenceTable<'db>lang_items: &'db LangItemstraits_in_scope: FxHashSet<TraitId>

The traits in scope, disregarding block modules. This is used for caching purposes.

result: InferenceResulttuple_field_accesses_rev: IndexSet<Tys<'db>, BuildHasherDefault<FxHasher>>return_ty: Ty<'db>

The return type of the function being inferred, the closure or async block if we’re currently within one.

We might consider using a nested inference context for checking closures so we can swap all shared things out at once.

return_coercion: Option<CoerceMany<'db, 'db>>

If Some, this stores coercion information for returned expressions. If None, this is in a context where return is inappropriate, such as a const expression.

resume_yield_tys: Option<(Ty<'db>, Ty<'db>)>

The resume type and the yield type, respectively, of the coroutine being inferred.

diverges: Divergesbreakables: Vec<BreakableContext<'db>>types: &'db DefaultAny<'db>inside_assignment: bool

Whether we are inside the pattern of a destructuring assignment.

deferred_cast_checks: Vec<CastCheck<'db>>current_captures: Vec<CapturedItemWithoutTy>current_capture_span_stack: Vec<MirSpan>

A stack that has an entry for each projection in the current capture.

For example, in a.b.c, we capture the spans of a, a.b, and a.b.c. We do that because sometimes we truncate projections (when a closure captures both a.b and a.b.c), and we want to provide accurate spans in this case.

current_closure: Option<InternedClosureId>closure_dependencies: FxHashMap<InternedClosureId, Vec<InternedClosureId>>

Stores the list of closure ids that need to be analyzed before this closure. See the comment on InferenceContext::sort_closures

deferred_closures: FxHashMap<InternedClosureId, Vec<(Ty<'db>, Ty<'db>, Vec<Ty<'db>>, ExprId)>>diagnostics: Diagnostics
Implementations

impl<'db> InferenceContext<'_, 'db>

fn place_of_expr(&mut self, tgt_expr: ExprId) -> Option<HirPlace>

fn path_place(&mut self, path: &Path, id: ExprOrPatId) -> Option<HirPlace>

Pushes the span into current_capture_span_stack, without clearing it first.

fn place_of_expr_without_adjust(&mut self, tgt_expr: ExprId) -> Option<HirPlace>

Changes current_capture_span_stack to contain the stack of spans for this expr.

fn push_capture(&mut self, place: HirPlace, kind: CaptureKind)

fn truncate_capture_spans( &self, capture: &mut CapturedItemWithoutTy, truncate_to: usize, )

fn ref_expr(&mut self, expr: ExprId, place: Option<HirPlace>)

fn add_capture(&mut self, place: HirPlace, kind: CaptureKind)

fn mutate_path_pat(&mut self, path: &Path, id: PatId)

fn mutate_expr(&mut self, expr: ExprId, place: Option<HirPlace>)

fn consume_expr(&mut self, expr: ExprId)

fn consume_place(&mut self, place: HirPlace)

fn walk_expr_with_adjust(&mut self, tgt_expr: ExprId, adjustment: &[Adjustment])

fn ref_capture_with_adjusts( &mut self, m: Mutability, tgt_expr: ExprId, rest: &[Adjustment], )

fn walk_expr(&mut self, tgt_expr: ExprId)

fn walk_expr_without_adjust(&mut self, tgt_expr: ExprId)

fn walk_pat(&mut self, result: &mut Option<CaptureKind>, pat: PatId)

fn walk_pat_inner( &mut self, p: PatId, update_result: &mut impl FnMut(CaptureKind), for_mut: BorrowKind, )

fn is_upvar(&self, place: &HirPlace) -> bool

fn is_ty_copy(&mut self, ty: Ty<'db>) -> bool

fn select_from_expr(&mut self, expr: ExprId)

fn restrict_precision_for_unsafe(&mut self)

fn adjust_for_move_closure(&mut self)

fn minimize_captures(&mut self)

fn consume_with_pat(&mut self, place: HirPlace, tgt_pat: PatId)

fn consume_exprs(&mut self, exprs: impl Iterator<Item = ExprId>)

fn closure_kind(&self) -> FnTrait

fn analyze_closure(&mut self, closure: InternedClosureId) -> FnTrait

fn strip_captures_ref_span(&mut self)

pub(crate) fn infer_closures(&mut self)

fn sort_closures( &mut self, ) -> Vec<(InternedClosureId, Vec<(Ty<'db>, Ty<'db>, Vec<Ty<'db>>, ExprId)>)>

We want to analyze some closures before others, to have a correct analysis:

We should analyze nested closures before the parent, since the parent should capture some of the things that its children captures.

If a closure calls another closure, we need to analyze the callee, to find out how we should capture it (e.g. by move for FnOnce)

These dependencies are collected in the main inference. We do a topological sort in this function. It will consume the deferred_closures field and return its content in a sorted vector.

pub(crate) fn add_current_closure_dependency(&mut self, dep: InternedClosureId)

impl<'db> InferenceContext<'_, 'db>

pub(super) fn infer_closure( &mut self, body: ExprId, args: &[PatId], ret_type: Option<TypeRefId>, arg_types: &[Option<TypeRefId>], closure_kind: ClosureKind, tgt_expr: ExprId, expected: &Expectation<'db>, ) -> Ty<'db>

fn fn_trait_kind_from_def_id(&self, trait_id: TraitId) -> Option<ClosureKind>

fn async_fn_trait_kind_from_def_id( &self, trait_id: TraitId, ) -> Option<ClosureKind>

fn deduce_closure_signature( &mut self, expected_ty: Ty<'db>, closure_kind: ClosureKind, ) -> (Option<PolyFnSig<'db>>, Option<ClosureKind>)

Given the expected type, figures out what it can about this closure we are about to type check:

fn deduce_closure_signature_from_predicates( &mut self, expected_ty: Ty<'db>, closure_kind: ClosureKind, predicates: impl DoubleEndedIterator<Item = Predicate<'db>>, ) -> (Option<PolyFnSig<'db>>, Option<ClosureKind>)

fn deduce_sig_from_projection( &mut self, closure_kind: ClosureKind, projection: PolyProjectionPredicate<'db>, ) -> Option<PolyFnSig<'db>>

Given a projection like “<F as Fn(X)>::Result == Y”, we can deduce everything we need to know about a closure or coroutine.

The cause_span should be the span that caused us to have this expected signature, or None if we can’t readily know that.

fn extract_sig_from_projection( &self, projection: PolyProjectionPredicate<'db>, ) -> Option<PolyFnSig<'db>>

Given an FnOnce::Output or AsyncFn::Output projection, extract the args and return type to infer a PolyFnSig for the closure.

fn extract_sig_from_projection_and_future_bound( &mut self, projection: PolyProjectionPredicate<'db>, ) -> Option<PolyFnSig<'db>>

When an async closure is passed to a function that has a “two-part” Fn and Future trait bound, like:

use std::future::Future;

fn not_exactly_an_async_closure<F, Fut>(_f: F)
where
    F: FnOnce(String, u32) -> Fut,
    Fut: Future<Output = i32>,
{}

The we want to be able to extract the signature to guide inference in the async closure. We will have two projection predicates registered in this case. First, we identify the FnOnce<Args, Output = ?Fut> bound, and if the output type is an inference variable ?Fut, we check if that is bounded by a Future<Output = Ty> projection.

This function is actually best-effort with the return type; if we don’t find a Future projection, we still will return arguments that we extracted from the FnOnce projection, and the output will be an unconstrained type variable instead.

fn sig_of_closure( &mut self, decl_inputs: &[Option<TypeRefId>], decl_output: Option<TypeRefId>, expected_sig: Option<PolyFnSig<'db>>, ) -> ClosureSignatures<'db>

fn sig_of_closure_no_expectation( &mut self, decl_inputs: &[Option<TypeRefId>], decl_output: Option<TypeRefId>, ) -> ClosureSignatures<'db>

If there is no expected signature, then we will convert the types that the user gave into a signature.

fn sig_of_closure_with_expectation( &mut self, decl_inputs: &[Option<TypeRefId>], decl_output: Option<TypeRefId>, expected_sig: PolyFnSig<'db>, ) -> ClosureSignatures<'db>

Invoked to compute the signature of a closure expression. This combines any user-provided type annotations (e.g., |x: u32| -> u32 { .. }) with the expected signature.

The approach is as follows:

Let S be the (higher-ranked) signature that we derive from the user’s annotations.

Let E be the (higher-ranked) signature that we derive from the expectations, if any.

If we have no expectation E, then the signature of the closure is S.

Otherwise, the signature of the closure is E. Moreover:

Skolemize the late-bound regions in E, yielding E'.

Instantiate all the late-bound regions bound in the closure within S with fresh (existential) variables, yielding S'

Require that E' = S'

We could use some kind of subtyping relationship here, I imagine, but equality is easier and works fine for our purposes.

The key intuition here is that the user’s types must be valid from “the inside” of the closure, but the expectation ultimately drives the overall signature.

Examples

fn with_closure<F>(_: F)
  where F: Fn(&u32) -> &u32 { .. }

with_closure(|x: &u32| { ... })

Here:

E would be fn(&u32) -> &u32.

S would be fn(&u32) -> ?T

E’ is &'!0 u32 -> &'!0 u32

S’ is &'?0 u32 -> ?T

S’ can be unified with E’ with ['?0 = '!0, ?T = &'!10 u32].

Arguments

expr_def_id: the LocalDefId of the closure expression

decl: the HIR declaration of the closure

body: the body of the closure

expected_sig: the expected signature (if any). Note that this is missing a binder: that is, there may be late-bound regions with depth 1, which are bound then by the closure.

fn sig_of_closure_with_mismatched_number_of_arguments( &mut self, decl_inputs: &[Option<TypeRefId>], decl_output: Option<TypeRefId>, ) -> ClosureSignatures<'db>

fn merge_supplied_sig_with_expectation( &mut self, decl_inputs: &[Option<TypeRefId>], decl_output: Option<TypeRefId>, expected_sigs: ClosureSignatures<'db>, ) -> InferResult<'db, ClosureSignatures<'db>>

Enforce the user’s types against the expectation. See sig_of_closure_with_expectation for details on the overall strategy.

fn supplied_sig_of_closure( &mut self, decl_inputs: &[Option<TypeRefId>], decl_output: Option<TypeRefId>, ) -> PolyFnSig<'db>

If there is no expected signature, then we will convert the types that the user gave into a signature.

Also, record this closure signature for later.

fn error_sig_of_closure( &mut self, decl_inputs: &[Option<TypeRefId>], decl_output: Option<TypeRefId>, ) -> PolyFnSig<'db>

Converts the types that the user supplied, in case that doing so should yield an error, but returns back a signature where all parameters are of type ty::Error.

fn closure_sigs(&self, bound_sig: PolyFnSig<'db>) -> ClosureSignatures<'db>

impl<'db> InferenceContext<'_, 'db>

pub(crate) fn coerce( &mut self, expr: ExprOrPatId, expr_ty: Ty<'db>, target: Ty<'db>, allow_two_phase: AllowTwoPhase, expr_is_read: ExprIsRead, ) -> RelateResult<'db, Ty<'db>>

Attempt to coerce an expression to a type, and return the adjusted type of the expression, if successful. Adjustments are only recorded if the coercion succeeded. The expressions must not have any preexisting adjustments.

fn try_find_coercion_lub( &mut self, exprs: &[ExprId], prev_ty: Ty<'db>, new: ExprId, new_ty: Ty<'db>, ) -> RelateResult<'db, Ty<'db>>

Given some expressions, their known unified type and another expression, tries to unify the types, potentially inserting coercions on any of the provided expressions and returns their LUB (aka “common supertype”).

This is really an internal helper. From outside the coercion module, you should instantiate a CoerceMany instance.

impl<'db> InferenceContext<'_, 'db>

pub(crate) fn infer_expr( &mut self, tgt_expr: ExprId, expected: &Expectation<'db>, is_read: ExprIsRead, ) -> Ty<'db>

pub(crate) fn infer_expr_no_expect( &mut self, tgt_expr: ExprId, is_read: ExprIsRead, ) -> Ty<'db>

pub(super) fn infer_expr_coerce( &mut self, expr: ExprId, expected: &Expectation<'db>, is_read: ExprIsRead, ) -> Ty<'db>

Infer type of expression with possibly implicit coerce to the expected type. Return the type after possible coercion.

pub(super) fn expr_guaranteed_to_constitute_read_for_never( &mut self, expr: ExprId, is_read: ExprIsRead, ) -> bool

Whether this expression constitutes a read of value of the type that it evaluates to.

This is used to determine if we should consider the block to diverge if the expression evaluates to !, and if we should insert a NeverToAny coercion for values of type !.

This function generally returns false if the expression is a place expression and the parent expression is the scrutinee of a match or the pointee of an & addr-of expression, since both of those parent expressions take a place and not a value.

fn pat_guaranteed_to_constitute_read_for_never(&self, pat: PatId) -> bool

Whether this pattern constitutes a read of value of the scrutinee that it is matching against. This is used to determine whether we should perform NeverToAny coercions.

fn contains_explicit_ref_binding(&self, pat: PatId) -> bool

Checks if the pattern contains any ref or ref mut bindings, and if yes whether it contains mutable or just immutables ones.

fn is_syntactic_place_expr(&self, expr: ExprId) -> bool

pub(crate) fn check_lhs_assignable(&self, lhs: ExprId)

fn infer_expr_coerce_never( &mut self, expr: ExprId, expected: &Expectation<'db>, is_read: ExprIsRead, ) -> Ty<'db>

fn infer_expr_inner( &mut self, tgt_expr: ExprId, expected: &Expectation<'db>, is_read: ExprIsRead, ) -> Ty<'db>

fn demand_scrutinee_type( &mut self, scrut: ExprId, contains_ref_bindings: bool, no_arms: bool, scrutinee_is_read: ExprIsRead, ) -> Ty<'db>

fn infer_expr_path( &mut self, path: &Path, id: ExprOrPatId, scope_id: ExprId, ) -> Ty<'db>

fn infer_unop_expr( &mut self, unop: UnaryOp, oprnd: ExprId, expected: &Expectation<'db>, expr: ExprId, ) -> Ty<'db>

fn infer_async_block( &mut self, tgt_expr: ExprId, statements: &[Statement], tail: &Option<ExprId>, ) -> Ty<'db>

pub(crate) fn lower_async_block_type_impl_trait( &mut self, inner_ty: Ty<'db>, tgt_expr: ExprId, ) -> Ty<'db>

pub(crate) fn write_fn_trait_method_resolution( &mut self, fn_x: FnTrait, derefed_callee: Ty<'db>, adjustments: &mut Vec<Adjustment>, callee_ty: Ty<'db>, params: &[Ty<'db>], tgt_expr: ExprId, )

fn infer_expr_array( &mut self, array: &Array, expected: &Expectation<'db>, ) -> Ty<'db>

pub(super) fn infer_return(&mut self, expr: ExprId)

fn infer_expr_return(&mut self, ret: ExprId, expr: Option<ExprId>) -> Ty<'db>

fn infer_expr_become(&mut self, expr: ExprId) -> Ty<'db>

fn infer_expr_box( &mut self, inner_expr: ExprId, expected: &Expectation<'db>, ) -> Ty<'db>

fn infer_block( &mut self, expr: ExprId, statements: &[Statement], tail: Option<ExprId>, label: Option<LabelId>, expected: &Expectation<'db>, ) -> Ty<'db>

fn lookup_field( &mut self, receiver_ty: Ty<'db>, name: &Name, ) -> Option<(Ty<'db>, Either<FieldId, TupleFieldId>, Vec<Adjustment>, bool)>

fn infer_field_access( &mut self, tgt_expr: ExprId, receiver: ExprId, name: &Name, expected: &Expectation<'db>, ) -> Ty<'db>

fn instantiate_erroneous_method( &mut self, def_id: FunctionId, ) -> MethodCallee<'db>

fn infer_call( &mut self, tgt_expr: ExprId, callee: ExprId, args: &[ExprId], expected: &Expectation<'db>, ) -> Ty<'db>

fn check_call( &mut self, tgt_expr: ExprId, args: &[ExprId], callee_ty: Ty<'db>, param_tys: &[Ty<'db>], ret_ty: Ty<'db>, indices_to_skip: &[u32], is_varargs: bool, expected: &Expectation<'db>, ) -> Ty<'db>

fn infer_method_call( &mut self, tgt_expr: ExprId, receiver: ExprId, args: &[ExprId], method_name: &Name, generic_args: Option<&HirGenericArgs>, expected: &Expectation<'db>, ) -> Ty<'db>

fn check_method_call( &mut self, tgt_expr: ExprId, args: &[ExprId], sig: FnSig<'db>, expected: &Expectation<'db>, ) -> Ty<'db>

pub(crate) fn check_call_arguments( &mut self, call_expr: ExprId, formal_input_tys: &[Ty<'db>], formal_output: Ty<'db>, expectation: &Expectation<'db>, provided_args: &[ExprId], skip_indices: &[u32], c_variadic: bool, )

Generic function that factors out common logic from function calls, method calls and overloaded operators.

fn register_obligations_for_call(&mut self, callable_ty: Ty<'db>)

fn check_legacy_const_generics( &mut self, callee: Ty<'db>, args: &[ExprId], ) -> Box<[u32]>

Returns the argument indices to skip.

pub(super) fn with_breakable_ctx<T>( &mut self, kind: BreakableKind, ty: Option<Ty<'db>>, label: Option<LabelId>, cb: impl FnOnce(&mut Self) -> T, ) -> (Option<Ty<'db>>, T)

impl<'db> InferenceContext<'_, 'db>

pub(super) fn type_inference_fallback(&mut self)

fn diverging_fallback_behavior(&self) -> DivergingFallbackBehavior

fn fallback_types(&mut self) -> bool

fn fallback_if_possible( &mut self, ty: Ty<'db>, diverging_fallback: &FxHashMap<Ty<'db>, Ty<'db>>, ) -> bool

fn calculate_diverging_fallback( &self, unresolved_variables: &[Ty<'db>], behavior: DivergingFallbackBehavior, ) -> FxHashMap<Ty<'db>, Ty<'db>>

The “diverging fallback” system is rather complicated. This is a result of our need to balance ‘do the right thing’ with backwards compatibility.

“Diverging” type variables are variables created when we coerce a ! type into an unbound type variable ?X. If they never wind up being constrained, the “right and natural” thing is that ?X should “fallback” to !. This means that e.g. an expression like Some(return) will ultimately wind up with a type like Option<!> (presuming it is not assigned or constrained to have some other type).

However, the fallback used to be () (before the ! type was added). Moreover, there are cases where the ! type ‘leaks out’ from dead code into type variables that affect live code. The most common case is something like this:

match foo() {
    22 => Default::default(), // call this type `?D`
    _ => return, // return has type `!`
} // call the type of this match `?M`

Here, coercing the type ! into ?M will create a diverging type variable ?X where ?X <: ?M. We also have that ?D <: ?M. If ?M winds up unconstrained, then ?X will fallback. If it falls back to !, then all the type variables will wind up equal to ! – this includes the type ?D (since ! doesn’t implement Default, we wind up a “trait not implemented” error in code like this). But since the original fallback was (), this code used to compile with ?D = (). This is somewhat surprising, since Default::default() on its own would give an error because the types are insufficiently constrained.

Our solution to this dilemma is to modify diverging variables so that they can either fallback to ! (the default) or to () (the backwards compatibility case). We decide which fallback to use based on whether there is a coercion pattern like this:

?Diverging -> ?V
?NonDiverging -> ?V
?V != ?NonDiverging

Here ?Diverging represents some diverging type variable and ?NonDiverging represents some non-diverging type variable. ?V can be any type variable (diverging or not), so long as it is not equal to ?NonDiverging.

Intuitively, what we are looking for is a case where a “non-diverging” type variable (like ?M in our example above) is coerced into some variable ?V that would otherwise fallback to !. In that case, we make ?V fallback to !, along with anything that would flow into ?V.

The algorithm we use:

Identify all variables that are coerced into by a diverging variable. Do this by iterating over each diverging, unsolved variable and finding all variables reachable from there. Call that set D.

Walk over all unsolved, non-diverging variables, and find any variable that has an edge into D.

fn create_coercion_graph(&self) -> Graph<(), ()>

Returns a graph whose nodes are (unresolved) inference variables and where an edge ?A -> ?B indicates that the variable ?A is coerced to ?B.

fn root_vid(&self, ty: Ty<'db>) -> Option<TyVid>

If ty is an unresolved type variable, returns its root vid.

impl<'db> InferenceContext<'_, 'db>

pub(crate) fn infer_mut_body(&mut self)

fn infer_mut_expr(&mut self, tgt_expr: ExprId, mutability: Mutability)

fn infer_mut_expr_without_adjust( &mut self, tgt_expr: ExprId, mutability: Mutability, )

fn infer_mut_not_expr_iter(&mut self, exprs: impl Iterator<Item = ExprId>)

fn pat_iter_bound_mutability( &self, pat: impl Iterator<Item = PatId>, ) -> Mutability

fn pat_bound_mutability(&self, pat: PatId) -> Mutability

Checks if the pat contains a ref mut binding. Such paths makes the context of bounded expressions mutable. For example in let (ref mut x0, ref x1) = *it; we need to use DerefMut for *it but in let (ref x0, ref x1) = *it; we should use Deref.

impl<'a, 'db> InferenceContext<'a, 'db>

pub(crate) fn infer_assign_op_expr( &mut self, expr: ExprId, op: ArithOp, lhs: ExprId, rhs: ExprId, ) -> Ty<'db>

Checks a a <op>= b

pub(crate) fn infer_binop_expr( &mut self, expr: ExprId, op: BinaryOp, lhs_expr: ExprId, rhs_expr: ExprId, ) -> Ty<'db>

Checks a potentially overloaded binary operator.

fn enforce_builtin_binop_types( &mut self, lhs_ty: Ty<'db>, rhs_ty: Ty<'db>, category: BinOpCategory, ) -> Ty<'db>

fn infer_overloaded_binop( &mut self, expr: ExprId, lhs_expr: ExprId, rhs_expr: ExprId, op: BinaryOp, ) -> (Ty<'db>, Ty<'db>, Ty<'db>)

pub(crate) fn infer_user_unop( &mut self, ex: ExprId, operand_ty: Ty<'db>, op: UnaryOp, ) -> Ty<'db>

fn lookup_op_method( &mut self, lhs_ty: Ty<'db>, opt_rhs: Option<(ExprId, Ty<'db>)>, (opname, trait_did): (Symbol, Option<TraitId>), ) -> Result<MethodCallee<'db>, Vec<NextSolverError<'db>>>

fn lang_item_for_bin_op(&self, op: BinaryOp) -> (Symbol, Option<TraitId>)

fn lang_item_for_unop(&self, op: UnaryOp) -> (Symbol, Option<TraitId>)

impl<'db> InferenceContext<'_, 'db>

pub(super) fn handle_opaque_type_uses(&mut self)

This takes all the opaque type uses during HIR typeck. It first computes the concrete hidden type by iterating over all defining uses.

A use during HIR typeck is defining if all non-lifetime arguments are unique generic parameters and the hidden type does not reference any inference variables.

It then uses these defining uses to guide inference for all other uses.

impl<'db> InferenceContext<'_, 'db>

fn compute_definition_site_hidden_types( &mut self, opaque_types: Vec<(OpaqueTypeKey<'db>, OpaqueHiddenType<'db>)>, )

fn consider_opaque_type_use( &self, opaque_type_key: OpaqueTypeKey<'db>, hidden_type: OpaqueHiddenType<'db>, ) -> UsageKind<'db>

impl<'db> InferenceContext<'_, 'db>

pub(super) fn infer_tuple_struct_pat_like( &mut self, path: Option<&Path>, expected: Ty<'db>, default_bm: BindingMode, id: PatId, ellipsis: Option<u32>, subs: &[PatId], decl: Option<DeclContext>, ) -> Ty<'db>

Infers type for tuple struct pattern or its corresponding assignee expression.

Ellipses found in the original pattern or expression must be filtered out.

pub(super) fn infer_record_pat_like( &mut self, path: Option<&Path>, expected: Ty<'db>, default_bm: BindingMode, id: PatId, subs: impl ExactSizeIterator<Item = (Name, PatId)>, decl: Option<DeclContext>, ) -> Ty<'db>

Infers type for record pattern or its corresponding assignee expression.

pub(super) fn infer_tuple_pat_like( &mut self, pat: PatId, expected: Ty<'db>, default_bm: BindingMode, ellipsis: Option<u32>, elements: &[PatId], decl: Option<DeclContext>, ) -> Ty<'db>

Infers type for tuple pattern or its corresponding assignee expression.

Ellipses found in the original pattern or expression must be filtered out.

pub(super) fn infer_top_pat( &mut self, pat: PatId, expected: Ty<'db>, decl: Option<DeclContext>, )

The resolver needs to be updated to the surrounding expression when inside assignment (because there, Pat::Path can refer to a variable).

fn infer_pat( &mut self, pat: PatId, expected: Ty<'db>, default_bm: BindingMode, decl: Option<DeclContext>, ) -> Ty<'db>

fn pat_ty_after_adjustment(&self, pat: PatId) -> Ty<'db>

fn infer_ref_pat( &mut self, inner_pat: PatId, mutability: Mutability, expected: Ty<'db>, default_bm: BindingMode, decl: Option<DeclContext>, ) -> Ty<'db>

fn infer_bind_pat( &mut self, pat: PatId, binding: BindingId, default_bm: BindingMode, subpat: Option<PatId>, expected: Ty<'db>, decl: Option<DeclContext>, ) -> Ty<'db>

fn infer_slice_pat( &mut self, expected: Ty<'db>, prefix: &[PatId], slice: Option<PatId>, suffix: &[PatId], default_bm: BindingMode, decl: Option<DeclContext>, ) -> Ty<'db>

fn infer_lit_pat(&mut self, expr: ExprId, expected: Ty<'db>) -> Ty<'db>

fn is_non_ref_pat(&mut self, body: &Body, pat: PatId) -> bool

fn try_resolve_slice_ty_to_array_ty( &mut self, before: &[PatId], suffix: &[PatId], slice: Option<PatId>, ) -> Option<Ty<'db>>

fn pat_is_irrefutable(&self, decl_ctxt: Option<DeclContext>) -> bool

Used to determine whether we can infer the expected type in the slice pattern to be of type array. This is only possible if we’re in an irrefutable pattern. If we were to allow this in refutable patterns we wouldn’t e.g. report ambiguity in the following situation:

   struct Zeroes;
   const ARR: [usize; 2] = [0; 2];
   const ARR2: [usize; 2] = [2; 2];

   impl Into<&'static [usize; 2]> for Zeroes {
       fn into(self) -> &'static [usize; 2] {
           &ARR
       }
   }

   impl Into<&'static [usize]> for Zeroes {
       fn into(self) -> &'static [usize] {
           &ARR2
       }
   }

   fn main() {
       let &[a, b]: &[usize] = Zeroes.into() else {
          ..
       };
   }

If we’re in an irrefutable pattern we prefer the array impl candidate given that the slice impl candidate would be rejected anyway (if no ambiguity existed).

impl<'db> InferenceContext<'_, 'db>

pub(super) fn infer_path( &mut self, path: &Path, id: ExprOrPatId, ) -> Option<Ty<'db>>

fn resolve_value_path( &mut self, path: &Path, id: ExprOrPatId, ) -> Option<ValuePathResolution<'db>>

pub(super) fn resolve_value_path_inner( &mut self, path: &Path, id: ExprOrPatId, no_diagnostics: bool, ) -> Option<(ValueNs, Option<GenericArgs<'db>>)>

fn add_required_obligations_for_value_path( &mut self, def: GenericDefId, subst: GenericArgs<'db>, )

fn resolve_trait_assoc_item( &mut self, trait_ref: TraitRef<'db>, segment: PathSegment<'_>, id: ExprOrPatId, ) -> Option<(ValueNs, GenericArgs<'db>)>

fn resolve_ty_assoc_item( &mut self, ty: Ty<'db>, name: &Name, id: ExprOrPatId, ) -> Option<(ValueNs, GenericArgs<'db>)>

fn resolve_enum_variant_on_ty( &mut self, ty: Ty<'db>, name: &Name, id: ExprOrPatId, ) -> Option<(ValueNs, GenericArgs<'db>)>

impl<'a, 'db> InferenceContext<'a, 'db>

pub(super) fn try_overloaded_deref( &self, base_ty: Ty<'db>, ) -> Option<InferOk<'db, MethodCallee<'db>>>

fn make_overloaded_place_return_type( &self, method: MethodCallee<'db>, ) -> Ty<'db>

For the overloaded place expressions (*x, x[3]), the trait returns a type of &T, but the actual type we assign to the expression is T. So this function just peels off the return type by one layer to yield T.

pub(super) fn lookup_derefing( &mut self, expr: ExprId, oprnd_expr: ExprId, oprnd_ty: Ty<'db>, ) -> Option<Ty<'db>>

Type-check *oprnd_expr with oprnd_expr type-checked already.

pub(super) fn lookup_indexing( &mut self, expr: ExprId, base_expr: ExprId, base_ty: Ty<'db>, idx_ty: Ty<'db>, ) -> Option<(Ty<'db>, Ty<'db>)>

Type-check *base_expr[index_expr] with base_expr and index_expr type-checked already.

fn try_index_step( expr: ExprId, base_expr: ExprId, autoderef: &mut GeneralAutoderef<'db, InferenceContextAutoderefCtx<'_, 'a, 'db>, Vec<(Ty<'db>, AutoderefKind)>>, index_ty: Ty<'db>, ) -> Option<(Ty<'db>, Ty<'db>)>

To type-check base_expr[index_expr], we progressively autoderef (and otherwise adjust) base_expr, looking for a type which either supports builtin indexing or overloaded indexing. This loop implements one step in that search; the autoderef loop is implemented by lookup_indexing.

pub(super) fn try_overloaded_place_op( &self, base_ty: Ty<'db>, opt_rhs_ty: Option<Ty<'db>>, op: PlaceOp, ) -> Option<InferOk<'db, MethodCallee<'db>>>

Try to resolve an overloaded place op. We only deal with the immutable variant here (Deref/Index). In some contexts we would need the mutable variant (DerefMut/IndexMut); those would be later converted by convert_place_derefs_to_mutable.

pub(super) fn try_mutable_overloaded_place_op( table: &InferenceTable<'db>, base_ty: Ty<'db>, opt_rhs_ty: Option<Ty<'db>>, op: PlaceOp, ) -> Option<InferOk<'db, MethodCallee<'db>>>

pub(super) fn convert_place_op_to_mutable( &mut self, op: PlaceOp, expr: ExprId, base_expr: ExprId, index_expr: Option<ExprId>, )

impl<'body, 'db> InferenceContext<'body, 'db>

fn new( db: &'db dyn HirDatabase, owner: DefWithBodyId, body: &'body Body, resolver: Resolver<'db>, ) -> Self

fn krate(&self) -> Crate

fn target_features(&self) -> (&TargetFeatures<'db>, TargetFeatureIsSafeInTarget)

fn set_tainted_by_errors(&mut self)

pub(crate) fn fixme_resolve_all_clone(&self) -> InferenceResult

Clones self and calls resolve_all() on it.

fn resolve_all(self) -> InferenceResult

fn collect_const(&mut self, id: ConstId, data: &ConstSignature)

fn collect_static(&mut self, data: &StaticSignature)

fn collect_fn(&mut self, func: FunctionId)

pub(crate) fn interner(&self) -> DbInterner<'db>

pub(crate) fn infcx(&self) -> &InferCtxt<'db>

fn infer_body(&mut self)

fn write_expr_ty(&mut self, expr: ExprId, ty: Ty<'db>)

pub(crate) fn write_expr_adj( &mut self, expr: ExprId, adjustments: Box<[Adjustment]>, )

fn write_pat_adj(&mut self, pat: PatId, adjustments: Box<[StoredTy]>)

pub(crate) fn write_method_resolution( &mut self, expr: ExprId, func: FunctionId, subst: GenericArgs<'db>, )

fn write_variant_resolution(&mut self, id: ExprOrPatId, variant: VariantId)

fn write_assoc_resolution( &mut self, id: ExprOrPatId, item: CandidateId, subs: GenericArgs<'db>, )

fn write_pat_ty(&mut self, pat: PatId, ty: Ty<'db>)

fn write_type_placeholder_ty(&mut self, type_ref: TypeRefId, ty: Ty<'db>)

fn write_binding_ty(&mut self, id: BindingId, ty: Ty<'db>)

pub(crate) fn push_diagnostic(&self, diagnostic: InferenceDiagnostic)

fn with_ty_lowering<R>( &mut self, store: &ExpressionStore, types_source: InferenceTyDiagnosticSource, lifetime_elision: LifetimeElisionKind<'db>, f: impl FnOnce(&mut TyLoweringContext<'db, '_>) -> R, ) -> R

fn with_body_ty_lowering<R>( &mut self, f: impl FnOnce(&mut TyLoweringContext<'db, '_>) -> R, ) -> R

fn make_ty( &mut self, type_ref: TypeRefId, store: &ExpressionStore, type_source: InferenceTyDiagnosticSource, lifetime_elision: LifetimeElisionKind<'db>, ) -> Ty<'db>

pub(crate) fn make_body_ty(&mut self, type_ref: TypeRefId) -> Ty<'db>

pub(crate) fn make_body_const( &mut self, const_ref: ConstRef, ty: Ty<'db>, ) -> Const<'db>

pub(crate) fn make_path_as_body_const( &mut self, path: &Path, ty: Ty<'db>, ) -> Const<'db>

fn err_ty(&self) -> Ty<'db>

pub(crate) fn make_body_lifetime( &mut self, lifetime_ref: LifetimeRefId, ) -> Region<'db>

fn insert_type_vars_shallow(&mut self, ty: Ty<'db>) -> Ty<'db>

Replaces Ty::Error by a new type var, so we can maybe still infer it.

fn insert_type_vars<T>(&mut self, ty: T) -> T
where T: TypeFoldable<DbInterner<'db>>,

fn unify(&mut self, ty1: Ty<'db>, ty2: Ty<'db>) -> bool

fn struct_tail_without_normalization(&mut self, ty: Ty<'db>) -> Ty<'db>

Attempts to returns the deeply last field of nested structures, but does not apply any normalization in its search. Returns the same type if input ty is not a structure at all.

fn struct_tail_with_normalize( &mut self, ty: Ty<'db>, normalize: impl FnMut(Ty<'db>) -> Ty<'db>, ) -> Ty<'db>

Returns the deeply last field of nested structures, or the same type if not a structure at all. Corresponds to the only possible unsized field, and its type can be used to determine unsizing strategy.

This is parameterized over the normalization strategy (i.e. how to handle <T as Trait>::Assoc and impl Trait); pass the identity function to indicate no normalization should take place.

fn process_user_written_ty(&mut self, ty: Ty<'db>) -> Ty<'db>

Whenever you lower a user-written type, you should call this.

fn process_remote_user_written_ty(&mut self, ty: Ty<'db>) -> Ty<'db>

The difference of this method from process_user_written_ty() is that this method doesn’t register a well-formed obligation, while process_user_written_ty() should (but doesn’t currently).

fn shallow_resolve(&self, ty: Ty<'db>) -> Ty<'db>

fn resolve_associated_type( &mut self, inner_ty: Ty<'db>, assoc_ty: Option<TypeAliasId>, ) -> Ty<'db>

fn demand_eqtype( &mut self, id: ExprOrPatId, expected: Ty<'db>, actual: Ty<'db>, ) -> Result<(), ()>

fn demand_eqtype_fixme_no_diag( &mut self, expected: Ty<'db>, actual: Ty<'db>, ) -> Result<(), ()>

fn demand_suptype(&mut self, expected: Ty<'db>, actual: Ty<'db>)

fn demand_coerce( &mut self, expr: ExprId, checked_ty: Ty<'db>, expected: Ty<'db>, allow_two_phase: AllowTwoPhase, expr_is_read: ExprIsRead, ) -> Ty<'db>

fn expr_ty(&self, expr: ExprId) -> Ty<'db>

fn expr_ty_after_adjustments(&self, e: ExprId) -> Ty<'db>

fn resolve_associated_type_with_params( &mut self, inner_ty: Ty<'db>, assoc_ty: Option<TypeAliasId>, params: &[GenericArg<'db>], ) -> Ty<'db>

fn resolve_variant( &mut self, node: ExprOrPatId, path: Option<&Path>, value_ns: bool, ) -> (Ty<'db>, Option<VariantId>)

fn resolve_variant_on_alias( &mut self, ty: Ty<'db>, unresolved: Option<usize>, path: &ModPath, ) -> (Ty<'db>, Option<VariantId>)

fn resolve_output_on(&self, trait_: TraitId) -> Option<TypeAliasId>

fn resolve_future_future_output(&self) -> Option<TypeAliasId>

fn resolve_boxed_box(&self) -> Option<AdtId>

fn resolve_range_full(&self) -> Option<AdtId>

fn has_new_range_feature(&self) -> bool

fn resolve_range(&self) -> Option<AdtId>

fn resolve_range_inclusive(&self) -> Option<AdtId>

fn resolve_range_from(&self) -> Option<AdtId>

fn resolve_range_to(&self) -> Option<AdtId>

fn resolve_range_to_inclusive(&self) -> Option<AdtId>

fn resolve_va_list(&self) -> Option<AdtId>

pub(crate) fn get_traits_in_scope( &self, ) -> Either<FxHashSet<TraitId>, &FxHashSet<TraitId>>

impl<'a, 'db> InferenceContext<'a, 'db>

pub(crate) fn confirm_method( &mut self, pick: &Pick<'db>, unadjusted_self_ty: Ty<'db>, expr: ExprId, generic_args: Option<&HirGenericArgs>, ) -> ConfirmResult<'db>

impl<'a, 'db> InferenceContext<'a, 'db>

pub(crate) fn lookup_method_including_private( &mut self, self_ty: Ty<'db>, name: Name, generic_args: Option<&HirGenericArgs>, receiver: ExprId, call_expr: ExprId, ) -> Result<(MethodCallee<'db>, bool), MethodError<'db>>

Performs method lookup. If lookup is successful, it will return the callee and store an appropriate adjustment for the self-expr. In some cases it may report an error (e.g., invoking the drop method).

pub(crate) fn lookup_probe( &self, method_name: Name, self_ty: Ty<'db>, ) -> Result<Pick<'db>, MethodError<'db>>

pub(crate) fn with_method_resolution<R>( &self, f: impl FnOnce(&MethodResolutionContext<'_, 'db>) -> R, ) -> R