hir_ty::infer::unify

Struct InferenceTable

pub(crate) struct InferenceTable<'db> {
    pub(crate) db: &'db dyn HirDatabase,
    pub(crate) param_env: ParamEnv<'db>,
    pub(crate) infer_ctxt: InferCtxt<'db>,
    pub(super) fulfillment_cx: FulfillmentCtxt<'db>,
    pub(super) diverging_type_vars: FxHashSet<Ty<'db>>,
}

Fields
db: &'db dyn HirDatabaseparam_env: ParamEnv<'db>infer_ctxt: InferCtxt<'db>fulfillment_cx: FulfillmentCtxt<'db>diverging_type_vars: FxHashSet<Ty<'db>>
Implementations

impl<'db> InferenceTable<'db>

pub(crate) fn autoderef( &self, base_ty: Ty<'db>, ) -> GeneralAutoderef<'db, DefaultAutoderefCtx<'_, 'db>, usize>

pub(crate) fn autoderef_with_tracking( &self, base_ty: Ty<'db>, ) -> GeneralAutoderef<'db, DefaultAutoderefCtx<'_, 'db>, Vec<(Ty<'db>, AutoderefKind)>>

impl<'db> InferenceTable<'db>

pub(crate) fn new( db: &'db dyn HirDatabase, trait_env: ParamEnv<'db>, krate: Crate, owner: Option<DefWithBodyId>, ) -> Self

Inside hir-ty you should use this for inference only, and always pass owner. Outside it, always pass owner = None.

pub(crate) fn interner(&self) -> DbInterner<'db>

pub(crate) fn type_is_copy_modulo_regions(&self, ty: Ty<'db>) -> bool

pub(crate) fn type_var_is_sized(&self, self_ty: TyVid) -> bool

pub(super) fn obligations_for_self_ty( &self, self_ty: TyVid, ) -> SmallVec<[Obligation<'db, Predicate<'db>>; 4]>

fn predicate_has_self_ty( &self, predicate: Predicate<'db>, expected_vid: TyVid, ) -> bool

fn type_matches_expected_vid(&self, expected_vid: TyVid, ty: Ty<'db>) -> bool

pub(super) fn set_diverging(&mut self, ty: Ty<'db>)

pub(crate) fn canonicalize<T>(&mut self, t: T) -> Canonical<DbInterner<'db>, T>
where T: TypeFoldable<DbInterner<'db>>,

pub(crate) fn normalize_alias_ty(&mut self, alias: Ty<'db>) -> Ty<'db>

pub(crate) fn next_ty_var(&self) -> Ty<'db>

pub(crate) fn next_const_var(&self) -> Const<'db>

pub(crate) fn next_int_var(&self) -> Ty<'db>

pub(crate) fn next_float_var(&self) -> Ty<'db>

pub(crate) fn new_maybe_never_var(&mut self) -> Ty<'db>

pub(crate) fn next_region_var(&self) -> Region<'db>

pub(crate) fn next_var_for_param(&self, id: GenericParamId) -> GenericArg<'db>

pub(crate) fn resolve_completely<T>(&mut self, value: T) -> T
where T: TypeFoldable<DbInterner<'db>>,

pub(crate) fn unify<T: ToTrace<'db>>(&mut self, ty1: T, ty2: T) -> bool

Unify two relatable values (e.g. Ty) and register new trait goals that arise from that.

pub(crate) fn try_unify<T: ToTrace<'db>>( &mut self, t1: T, t2: T, ) -> InferResult<'db, ()>

Unify two relatable values (e.g. Ty) and return new trait goals arising from it, so the caller needs to deal with them.

pub(crate) fn at<'a>(&'a self, cause: &'a ObligationCause) -> At<'a, 'db>

pub(crate) fn shallow_resolve(&self, ty: Ty<'db>) -> Ty<'db>

pub(crate) fn resolve_vars_with_obligations<T>(&mut self, t: T) -> T
where T: TypeFoldable<DbInterner<'db>>,

pub(crate) fn fresh_args_for_item(&self, def: SolverDefId) -> GenericArgs<'db>

Create a GenericArgs full of infer vars for def.

pub(crate) fn try_structurally_resolve_type(&mut self, ty: Ty<'db>) -> Ty<'db>

Try to resolve ty to a structural type, normalizing aliases.

In case there is still ambiguity, the returned type may be an inference variable. This is different from structurally_resolve_type which errors in this case.

pub(crate) fn structurally_resolve_type(&mut self, ty: Ty<'db>) -> Ty<'db>

pub(crate) fn snapshot(&mut self) -> InferenceTableSnapshot<'db>

pub(crate) fn rollback_to(&mut self, snapshot: InferenceTableSnapshot<'db>)

pub(crate) fn commit_if_ok<T, E>( &mut self, f: impl FnOnce(&mut InferenceTable<'db>) -> Result<T, E>, ) -> Result<T, E>

pub(crate) fn try_obligation( &mut self, predicate: Predicate<'db>, ) -> NextTraitSolveResult

Checks an obligation without registering it. Useful mostly to check whether a trait might be implemented before deciding to ‘lock in’ the choice (during e.g. method resolution or deref).

pub(crate) fn register_obligation(&mut self, predicate: Predicate<'db>)

fn register_obligation_in_env(&mut self, goal: Goal<'db, Predicate<'db>>)

pub(crate) fn register_infer_ok<T>(&mut self, infer_ok: InferOk<'db, T>) -> T

pub(crate) fn select_obligations_where_possible(&mut self)

pub(super) fn register_predicate( &mut self, obligation: Obligation<'db, Predicate<'db>>, )

pub(crate) fn register_predicates<I>(&mut self, obligations: I)
where I: IntoIterator<Item = Obligation<'db, Predicate<'db>>>,

pub(crate) fn register_wf_obligation( &mut self, term: Term<'db>, cause: ObligationCause, )

checking later, during regionck, that arg is well-formed.

pub(crate) fn add_wf_bounds(&mut self, args: GenericArgs<'db>)

Registers obligations that all args are well-formed.

pub(crate) fn callable_sig( &mut self, ty: Ty<'db>, num_args: usize, ) -> Option<(Option<FnTrait>, Vec<Ty<'db>>, Ty<'db>)>

fn callable_sig_from_fn_trait( &mut self, ty: Ty<'db>, num_args: usize, ) -> Option<(FnTrait, Vec<Ty<'db>>, Ty<'db>)>

pub(super) fn insert_type_vars<T>(&mut self, ty: T) -> T
where T: TypeFoldable<DbInterner<'db>>,

pub(super) fn insert_type_vars_shallow(&mut self, ty: Ty<'db>) -> Ty<'db>

Replaces Ty::Error by a new type var, so we can maybe still infer it.

pub(crate) fn process_user_written_ty(&mut self, ty: Ty<'db>) -> Ty<'db>

Whenever you lower a user-written type, you should call this.

pub(crate) fn process_remote_user_written_ty(&mut self, ty: Ty<'db>) -> Ty<'db>

The difference of this method from process_user_written_ty() is that this method doesn’t register a well-formed obligation, while process_user_written_ty() should (but doesn’t currently).

pub(super) fn insert_const_vars_shallow(&mut self, c: Const<'db>) -> Const<'db>

Replaces ConstScalar::Unknown by a new type var, so we can maybe still infer it.

pub(crate) fn is_sized(&mut self, ty: Ty<'db>) -> bool

Check if given type is Sized or not

impl<'db> InferenceTable<'db>

pub(crate) fn lookup_method_for_operator( &self, cause: ObligationCause, method_name: Symbol, trait_def_id: TraitId, self_ty: Ty<'db>, opt_rhs_ty: Option<Ty<'db>>, treat_opaques: TreatNotYetDefinedOpaques, ) -> Option<InferOk<'db, MethodCallee<'db>>>

lookup_method_in_trait is used for overloaded operators. It does a very narrow slice of what the normal probe/confirm path does. In particular, it doesn’t really do any probing: it simply constructs an obligation for a particular trait with the given self type and checks whether that trait is implemented.