hir_ty

Module inhabitedness

Type inhabitedness logic.

Structs

UninhabitedFrom 🔒

VisiblyUninhabited 🔒

Constants

BREAK_VISIBLY_UNINHABITED 🔒

CONTINUE_OPAQUELY_INHABITED 🔒

Functions

is_enum_variant_uninhabited_from 🔒

Checks whether a variant is visibly uninhabited from a particular module.

is_ty_uninhabited_from 🔒

Checks whether a type is visibly uninhabited from a particular module.