Type Alias hir_ty::layout::Layout

pub type Layout = LayoutS<RustcFieldIdx, RustcEnumVariantIdx>;

Aliased Type

struct Layout {
    pub fields: FieldsShape<RustcFieldIdx>,
    pub variants: Variants<RustcFieldIdx, RustcEnumVariantIdx>,
    pub abi: Abi,
    pub largest_niche: Option<Niche>,
    pub align: AbiAndPrefAlign,
    pub size: Size,
    pub max_repr_align: Option<Align>,
    pub unadjusted_abi_align: Align,
}

Fields
fields: FieldsShape<RustcFieldIdx>

Says where the fields are located within the layout.

variants: Variants<RustcFieldIdx, RustcEnumVariantIdx>

Encodes information about multi-variant layouts. Even with Multiple variants, a layout still has its own fields! Those are then shared between all variants. One of them will be the discriminant, but e.g. coroutines can have more.

To access all fields of this layout, both fields and the fields of the active variant must be taken into account.

abi: Abi

The abi defines how this data is passed between functions, and it defines value restrictions via valid_range.

Note that this is entirely orthogonal to the recursive structure defined by variants and fields; for example, ManuallyDrop<Result<isize, isize>> has Abi::ScalarPair! So, even with non-Aggregate abi, fields and variants have to be taken into account to find all fields of this layout.

largest_niche: Option<Niche>

The leaf scalar with the largest number of invalid values (i.e. outside of its valid_range), if it exists.

align: AbiAndPrefAlignsize: Sizemax_repr_align: Option<Align>

The largest alignment explicitly requested with repr(align) on this type or any field. Only used on i686-windows, where the argument passing ABI is different when alignment is requested, even if the requested alignment is equal to the natural alignment.

unadjusted_abi_align: Align

The alignment the type would have, ignoring any repr(align) but including repr(packed). Only used on aarch64-linux, where the argument passing ABI ignores the requested alignment in some cases.