Type Alias hir_ty::layout::TagEncoding

pub type TagEncoding = TagEncoding<RustcEnumVariantIdx>;

Aliased Type

enum TagEncoding {
    Direct,
    Niche {
        untagged_variant: RustcEnumVariantIdx,
        niche_variants: RangeInclusive<RustcEnumVariantIdx>,
        niche_start: u128,
    },
}

Variants

Direct

The tag directly stores the discriminant, but possibly with a smaller layout (so converting the tag to the discriminant can require sign extension).

Niche

Niche (values invalid for a type) encoding the discriminant: Discriminant and variant index coincide. The variant untagged_variant contains a niche at an arbitrary offset (field tag_field of the enum), which for a variant with discriminant d is set to (d - niche_variants.start).wrapping_add(niche_start).

For example, Option<(usize, &T)> is represented such that None has a null pointer for the second tuple field, and Some is the identity function (with a non-null reference).

Fields

untagged_variant: RustcEnumVariantIdx

niche_variants: RangeInclusive<RustcEnumVariantIdx>

niche_start: u128