hir_ty

Module lower

Methods for lowering the HIR to types. There are two main cases here:

Lowering a type reference like &usize or Option<foo::bar::Baz> to a type: The entry point for this is TyLoweringContext::lower_ty.

Building the type for an item: This happens through the ty query.

This usually involves resolving names, collecting generic arguments etc.

Modules

diagnostics ğŸ”’

This files contains the declaration of diagnostics kinds for ty and path lowering.

path ğŸ”’

A wrapper around TyLoweringContext specifically for lowering paths.

Structs

GenericDefaults

GenericPredicates

ImplTrait

ImplTraitLoweringState ğŸ”’

ImplTraits

PathDiagnosticCallbackData ğŸ”’

SupertraitsInfo ğŸ”’

TyLoweringContext

TypeAliasBounds ğŸ”’

Enums

GenericPredicateSource ğŸ”’

ImplTraitLoweringMode ğŸ”’

LifetimeElisionKind

TyDefId

TypeParamAssocTypeShorthandError ğŸ”’

ValueTyDefId

Functions

associated_ty_item_bounds ğŸ”’

associated_type_by_name_including_super_traits ğŸ”’

callable_item_signature ğŸ”’

Build the signature of a callable item (function, struct or enum variant).

const_param_ty_query ğŸ”’

const_param_ty_with_diagnostics ğŸ”’

create_diagnostics ğŸ”’

dyn_trait_dummy_self ğŸ”’

field_types_query ğŸ”’

field_types_with_diagnostics_query ğŸ”’

Build the type of all specific fields of a struct or enum variant.

fn_sig_for_enum_variant_constructor ğŸ”’

fn_sig_for_fn ğŸ”’

fn_sig_for_struct_constructor ğŸ”’

generic_defaults_query ğŸ”’

generic_defaults_with_diagnostics_cycle_result ğŸ”’

generic_defaults_with_diagnostics_query ğŸ”’

Resolve the default type params from generics.

generic_predicates ğŸ”’

Resolve the where clause(s) of an item with generics, with a given filter

generic_predicates_cycle_result ğŸ”’

A cycle can occur from malformed code.

impl_self_ty_query ğŸ”’

impl_self_ty_with_diagnostics ğŸ”’

impl_trait_query ğŸ”’

impl_trait_with_diagnostics ğŸ”’

lower_mutability ğŸ”’

param_env_from_predicates ğŸ”’

push_const_arg_has_type_predicates ğŸ”’

replace_dummy_self_with_error ğŸ”’

resolve_type_param_assoc_type_shorthand ğŸ”’

Predicates for param_id of the form P: SomeTrait. If assoc_name is provided, only return predicates referencing traits that have an associated type of that name.

resolve_type_param_assoc_type_shorthand_cycle_result ğŸ”’

trait_environment ğŸ”’

trait_environment_for_body_query ğŸ”’

ty_query ğŸ”’

Build the declared type of an item. This depends on the namespace; e.g. for struct Foo(usize), we have two types: The type of the struct itself, and the constructor function (usize) -> Foo which lives in the values namespace.

type_alias_bounds ğŸ”’

type_alias_bounds_with_diagnostics ğŸ”’

type_alias_self_bounds ğŸ”’

type_for_adt ğŸ”’

type_for_const ğŸ”’

Build the declared type of a const.

type_for_enum_variant_constructor ğŸ”’

Build the type of a tuple enum variant constructor.

type_for_fn ğŸ”’

Build the declared type of a function. This should not need to look at the function body.

type_for_static ğŸ”’

Build the declared type of a static.

type_for_struct_constructor ğŸ”’

Build the type of a tuple struct constructor.

type_for_type_alias_with_diagnostics ğŸ”’

unknown_const ğŸ”’

value_ty ğŸ”’

Type Aliases

Diagnostics ğŸ”’

ImplTraitIdx