hir_ty::lower

Function all_supertraits_trait_refs

pub(crate) fn all_supertraits_trait_refs(
    db: &dyn HirDatabase,
    trait_: TraitId,
) -> impl ExactSizeIterator<Item = EarlyBinder<'_, TraitRef<'_>>>

During lowering, elaborating supertraits can cause cycles. To avoid that, we have a separate query to only collect supertraits.

Technically, it is possible to avoid even more cycles by only collecting the TraitId of supertraits without their args. However rustc doesn’t do that, so we don’t either.