hir_ty::lower

Function resolve_type_param_assoc_type_shorthand

fn resolve_type_param_assoc_type_shorthand<'db>(
    db: &'db dyn HirDatabase,
    def: GenericDefId,
    param: TypeParamId,
    assoc_name: Name,
) -> &'db Result<StoredEarlyBinder<(TypeAliasId, StoredGenericArgs)>, TypeParamAssocTypeShorthandError>

Predicates for param_id of the form P: SomeTrait. If assoc_name is provided, only return predicates referencing traits that have an associated type of that name.

This query exists only to be used when resolving short-hand associated types like T::Item.

See the analogous query in rustc and its comment: https://github.com/rust-lang/rust/blob/9150f844e2624eb013ec78ca08c1d416e6644026/src/librustc_typeck/astconv.rs#L46

This is a query mostly to handle cycles somewhat gracefully; e.g. the following bounds are disallowed: T: Foo<U::Item>, U: Foo<T::Item>, but these are fine: T: Foo<U::Item>, U: Foo<()>.