hir_ty

Module method_resolution

This module is concerned with finding methods that a given type provides. For details about how this works in rustc, see the method lookup page in the rustc guide and the corresponding code mostly in rustc_hir_analysis/check/method/probe.rs.

Modules

confirm 🔒

Confirmation step of method selection, meaning ensuring the selected candidate is valid and registering all obligations.

probe 🔒

Candidate assembly and selection in method resolution - where we enumerate all candidates and choose the best one (or, in some IDE scenarios, just enumerate them all).

Structs

Candidate

CandidateStep

CandidateWithPrivate

InherentImpls

MethodCallee 🔒

MethodResolutionContext

MethodResolutionUnstableFeatures

OneTraitImpls 🔒

OneTraitImplsBuilder 🔒

Pick

TraitImpls

Enums

CandidateId

CandidateKind

CandidateSource

MethodError

Mode

PickKind

TreatNotYetDefinedOpaques 🔒

Used by FnCtxt::lookup_method_for_operator with -Znext-solver.

Functions

crates_containing_incoherent_inherent_impls 🔒

find_matching_impl 🔒

is_dyn_method

Checks if the self parameter of Trait method is the dyn Trait and we should call the method using the vtable.

lookup_impl_assoc_item_for_trait_ref 🔒

lookup_impl_const

lookup_impl_method_query 🔒

Looks up the impl method that actually runs for the trait method func.

simplified_type_module

with_incoherent_inherent_impls