Function hir_ty::method_resolution::is_dyn_method

pub fn is_dyn_method(
    db: &dyn HirDatabase,
    _env: Arc<TraitEnvironment>,
    func: FunctionId,
    fn_subst: Substitution
) -> Option<usize>

Checks if the self parameter of Trait method is the dyn Trait and we should call the method using the vtable.