hir_ty::method_resolution

Module probe

Candidate assembly and selection in method resolution - where we enumerate all candidates and choose the best one (or, in some IDE scenarios, just enumerate them all).

Structs

Candidate

CandidateStep

CandidateWithPrivate

MethodAutoderefBadTy 🔒

MethodAutoderefStepsResult 🔒

Pick

PickConstraintsForShadowed 🔒

Criteria to apply when searching for a given Pick. This is used during the search for potentially shadowed methods to ensure we don’t search more candidates than strictly necessary.

ProbeAllChoice 🔒

ProbeContext 🔒

ProbeForNameChoice 🔒

Enums

AutorefOrPtrAdjustment

When adjusting a receiver we often want to do one of

CandidateKind

Mode

PickKind

ProbeResult 🔒

Traits

ProbeChoice 🔒

Functions

is_relevant_kind_for_mode 🔒

Determine if the given associated item type is relevant in the current context.

Type Aliases

PickResult 🔒