Module hir_ty::mir

MIR definitions and implementation

Structs

BasicBlock

BorrowckResult

Evaluator

Local

MirBody

Place

Statement

SwitchTargets

Enums

AggregateKind

BinOp

BorrowKind

CastKind

MirEvalError

MirLowerError

MirSpan

MutabilityReason

Stores spans which implies that the local should be mutable.

Operand

An operand in MIR represents a “value” in Rust, the definition of which is undecided and part of the memory model. One proposal for a definition of values can be found on UCG.

ProjectionElem

Rvalue

StatementKind

Terminator

UnOp

Functions

borrowck_query

interpret_mir

lower_to_mir

mir_body_for_closure_query

mir_body_query

mir_body_recover

pad16

Type Definitions

BasicBlockId

LocalId