hir_ty::mir

Module borrowck

MIR borrow checker, which is used in diagnostics like unused_mut

Structs

BorrowRegion

BorrowckResult

MovedOutOfRef

PartiallyMoved

Enums

MutabilityReason

Stores spans which implies that the local should be mutable.

ProjectionCase 🔒

Functions

all_mir_bodies 🔒

borrow_regions 🔒

borrowck_query

ever_initialized_map 🔒

Returns a map from basic blocks to the set of locals that might be ever initialized before the start of the block. Only StorageDead can remove something from this map, and we ignore Uninit and drop and similar after initialization.

make_fetch_closure_field 🔒

moved_out_of_ref 🔒

mutability_of_locals 🔒

partially_moved 🔒

place_case 🔒

push_mut_span 🔒

record_usage 🔒

record_usage_for_operand 🔒