hir_ty::mir::eval

Struct Evaluator

pub struct Evaluator<'db> {

    db: &'db dyn HirDatabase,
    param_env: ParamEnvAndCrate<'db>,
    target_data_layout: Arc<TargetDataLayout>,
    stack: Vec<u8>,
    heap: Vec<u8>,
    code_stack: Vec<StackFrame>,
    static_locations: FxHashMap<StaticId, Address>,
    vtable_map: VTableMap<'db>,
    thread_local_storage: TlsData,
    random_state: Rand64,
    stdout: Vec<u8>,
    stderr: Vec<u8>,
    layout_cache: RefCell<FxHashMap<Ty<'db>, Arc<Layout>>>,
    projected_ty_cache: RefCell<FxHashMap<(Ty<'db>, ProjectionElem<LocalId>), Ty<'db>>>,
    not_special_fn_cache: RefCell<FxHashSet<FunctionId>>,
    mir_or_dyn_index_cache: RefCell<FxHashMap<(FunctionId, GenericArgs<'db>), MirOrDynIndex>>,
    unused_locals_store: RefCell<FxHashMap<DefWithBodyId, Vec<Locals>>>,
    cached_ptr_size: usize,
    cached_fn_trait_func: Option<FunctionId>,
    cached_fn_mut_trait_func: Option<FunctionId>,
    cached_fn_once_trait_func: Option<FunctionId>,
    crate_id: Crate,
    assert_placeholder_ty_is_unused: bool,
    execution_limit: usize,
    stack_depth_limit: usize,
    memory_limit: usize,
    infcx: InferCtxt<'db>,

}

Fields
db: &'db dyn HirDatabaseparam_env: ParamEnvAndCrate<'db>target_data_layout: Arc<TargetDataLayout>stack: Vec<u8>heap: Vec<u8>code_stack: Vec<StackFrame>static_locations: FxHashMap<StaticId, Address>

Stores the global location of the statics. We const evaluate every static first time we need it and see it’s missing, then we add it to this to reuse.

vtable_map: VTableMap<'db>

We don’t really have function pointers, i.e. pointers to some assembly instructions that we can run. Instead, we store the type as an interned id in place of function and vtable pointers, and we recover back the type at the time of use.

thread_local_storage: TlsDatarandom_state: Rand64stdout: Vec<u8>stderr: Vec<u8>layout_cache: RefCell<FxHashMap<Ty<'db>, Arc<Layout>>>projected_ty_cache: RefCell<FxHashMap<(Ty<'db>, ProjectionElem<LocalId>), Ty<'db>>>not_special_fn_cache: RefCell<FxHashSet<FunctionId>>mir_or_dyn_index_cache: RefCell<FxHashMap<(FunctionId, GenericArgs<'db>), MirOrDynIndex>>unused_locals_store: RefCell<FxHashMap<DefWithBodyId, Vec<Locals>>>

Constantly dropping and creating Locals is very costly. We store old locals that we normally want to drop here, to reuse their allocations later.

cached_ptr_size: usizecached_fn_trait_func: Option<FunctionId>cached_fn_mut_trait_func: Option<FunctionId>cached_fn_once_trait_func: Option<FunctionId>crate_id: Crateassert_placeholder_ty_is_unused: boolexecution_limit: usize

A general limit on execution, to prevent non terminating programs from breaking r-a main process

stack_depth_limit: usize

An additional limit on stack depth, to prevent stack overflow

memory_limit: usize

Maximum count of bytes that heap and stack can grow

infcx: InferCtxt<'db>
Implementations

impl<'db> Evaluator<'db>

fn detect_simd_ty(&self, ty: Ty<'db>) -> Result<(usize, Ty<'db>), MirEvalError>

pub(super) fn exec_simd_intrinsic( &mut self, name: &str, args: &[IntervalAndTy<'db>], _generic_args: GenericArgs<'db>, destination: Interval, _locals: &Locals, _span: MirSpan, ) -> Result<(), MirEvalError>

impl<'db> Evaluator<'db>

pub(super) fn detect_and_exec_special_function( &mut self, def: FunctionId, args: &[IntervalAndTy<'db>], generic_args: GenericArgs<'db>, locals: &Locals, destination: Interval, span: MirSpan, ) -> Result<bool, MirEvalError>

pub(super) fn detect_and_redirect_special_function( &mut self, def: FunctionId, ) -> Result<Option<FunctionId>, MirEvalError>

fn exec_clone( &mut self, def: FunctionId, args: &[IntervalAndTy<'db>], self_ty: Ty<'db>, locals: &Locals, destination: Interval, span: MirSpan, ) -> Result<(), MirEvalError>

Clone has special impls for tuples and function pointers

fn exec_clone_for_fields( &mut self, ty_iter: impl Iterator<Item = Ty<'db>>, layout: Arc<Layout>, addr: Address, def: FunctionId, locals: &Locals, destination: Interval, span: MirSpan, ) -> Result<(), MirEvalError>

fn exec_alloc_fn( &mut self, alloc_fn: AttrFlags, args: &[IntervalAndTy<'db>], destination: Interval, ) -> Result<(), MirEvalError>

fn detect_lang_function(&self, def: FunctionId) -> Option<EvalLangItem>

fn exec_lang_item( &mut self, it: EvalLangItem, generic_args: GenericArgs<'db>, args: &[IntervalAndTy<'db>], locals: &Locals, span: MirSpan, ) -> Result<Vec<u8>, MirEvalError>

fn exec_syscall( &mut self, id: i64, args: &[IntervalAndTy<'db>], destination: Interval, _locals: &Locals, _span: MirSpan, ) -> Result<(), MirEvalError>

fn exec_extern_c( &mut self, as_str: &str, args: &[IntervalAndTy<'db>], _generic_args: GenericArgs<'db>, destination: Interval, locals: &Locals, span: MirSpan, ) -> Result<(), MirEvalError>

fn exec_intrinsic( &mut self, name: &str, args: &[IntervalAndTy<'db>], generic_args: GenericArgs<'db>, destination: Interval, locals: &Locals, span: MirSpan, needs_override: bool, ) -> Result<bool, MirEvalError>

fn size_align_of_unsized( &mut self, ty: Ty<'db>, metadata: Interval, locals: &Locals, ) -> Result<(usize, usize), MirEvalError>

fn exec_atomic_intrinsic( &mut self, name: &str, args: &[IntervalAndTy<'db>], generic_args: GenericArgs<'db>, destination: Interval, locals: &Locals, _span: MirSpan, ) -> Result<(), MirEvalError>

impl<'db> Evaluator<'db>

pub fn new( db: &'db dyn HirDatabase, owner: DefWithBodyId, assert_placeholder_ty_is_unused: bool, trait_env: Option<ParamEnvAndCrate<'db>>, ) -> Result<Evaluator<'db>, MirEvalError>

fn interner(&self) -> DbInterner<'db>

fn lang_items(&self) -> &'db LangItems

fn place_addr( &self, p: &Place, locals: &Locals, ) -> Result<Address, MirEvalError>

fn place_interval( &self, p: &Place, locals: &Locals, ) -> Result<Interval, MirEvalError>

fn ptr_size(&self) -> usize

fn projected_ty(&self, ty: Ty<'db>, proj: ProjectionElem<LocalId>) -> Ty<'db>

fn place_addr_and_ty_and_metadata<'a>( &'a self, p: &Place, locals: &'a Locals, ) -> Result<(Address, Ty<'db>, Option<IntervalOrOwned>), MirEvalError>

fn layout(&self, ty: Ty<'db>) -> Result<Arc<Layout>, MirEvalError>

fn layout_adt( &self, adt: AdtId, subst: GenericArgs<'db>, ) -> Result<Arc<Layout>, MirEvalError>

fn place_ty<'a>( &'a self, p: &Place, locals: &'a Locals, ) -> Result<Ty<'db>, MirEvalError>

fn operand_ty( &self, o: &Operand, locals: &Locals, ) -> Result<Ty<'db>, MirEvalError>

fn operand_ty_and_eval( &mut self, o: &Operand, locals: &mut Locals, ) -> Result<IntervalAndTy<'db>, MirEvalError>

fn interpret_mir( &mut self, body: Arc<MirBody>, args: impl Iterator<Item = IntervalOrOwned>, ) -> Result<Interval, MirEvalError>

fn fill_locals_for_body( &mut self, body: &MirBody, locals: &mut Locals, args: impl Iterator<Item = IntervalOrOwned>, ) -> Result<(), MirEvalError>

fn create_locals_for_body( &mut self, body: &Arc<MirBody>, destination: Option<Interval>, ) -> Result<(Locals, usize), MirEvalError>

fn eval_rvalue( &mut self, r: &Rvalue, locals: &mut Locals, ) -> Result<IntervalOrOwned, MirEvalError>

fn compute_discriminant( &self, ty: Ty<'db>, bytes: &[u8], ) -> Result<i128, MirEvalError>

fn coerce_unsized_look_through_fields<T>( &self, ty: Ty<'db>, goal: impl Fn(TyKind<'db>) -> Option<T>, ) -> Result<T, MirEvalError>

fn coerce_unsized( &mut self, addr: Interval, current_ty: Ty<'db>, target_ty: Ty<'db>, ) -> Result<IntervalOrOwned, MirEvalError>

fn unsizing_ptr_from_addr( &mut self, target_ty: Ty<'db>, current_ty: Ty<'db>, addr: Interval, ) -> Result<IntervalOrOwned, MirEvalError>

Adds metadata to the address and create the fat pointer result of the unsizing operation.

fn layout_of_variant( &mut self, it: VariantId, subst: GenericArgs<'db>, locals: &Locals, ) -> Result<(usize, Arc<Layout>, Option<(usize, usize, i128)>), MirEvalError>

fn construct_with_layout( &mut self, size: usize, variant_layout: &Layout, tag: Option<(usize, usize, i128)>, values: impl Iterator<Item = IntervalOrOwned>, ) -> Result<Vec<u8>, MirEvalError>

fn eval_operand( &mut self, it: &Operand, locals: &mut Locals, ) -> Result<Interval, MirEvalError>

fn allocate_const_in_heap( &mut self, locals: &Locals, konst: Const<'db>, ) -> Result<Interval, MirEvalError>

fn eval_place( &mut self, p: &Place, locals: &Locals, ) -> Result<Interval, MirEvalError>

fn read_memory(&self, addr: Address, size: usize) -> Result<&[u8], MirEvalError>

fn write_memory_using_ref( &mut self, addr: Address, size: usize, ) -> Result<&mut [u8], MirEvalError>

fn write_memory(&mut self, addr: Address, r: &[u8]) -> Result<(), MirEvalError>

fn copy_from_interval_or_owned( &mut self, addr: Address, r: IntervalOrOwned, ) -> Result<(), MirEvalError>

fn copy_from_interval( &mut self, addr: Address, r: Interval, ) -> Result<(), MirEvalError>

fn size_align_of( &self, ty: Ty<'db>, locals: &Locals, ) -> Result<Option<(usize, usize)>, MirEvalError>

fn size_of_sized( &self, ty: Ty<'db>, locals: &Locals, what: &'static str, ) -> Result<usize, MirEvalError>

A version of self.size_of which returns error if the type is unsized. what argument should be something that complete this: error: type {ty} was unsized. {what} should be sized

fn size_align_of_sized( &self, ty: Ty<'db>, locals: &Locals, what: &'static str, ) -> Result<(usize, usize), MirEvalError>

A version of self.size_align_of which returns error if the type is unsized. what argument should be something that complete this: error: type {ty} was unsized. {what} should be sized

fn heap_allocate( &mut self, size: usize, align: usize, ) -> Result<Address, MirEvalError>

fn detect_fn_trait(&self, def: FunctionId) -> Option<FnTrait>

fn create_memory_map( &self, bytes: &[u8], ty: Ty<'db>, locals: &Locals, ) -> Result<ComplexMemoryMap<'db>, MirEvalError>

fn patch_addresses( &mut self, patch_map: &FxHashMap<usize, usize>, ty_of_bytes: impl Fn(&[u8]) -> Result<Ty<'db>, MirEvalError> + Copy, addr: Address, ty: Ty<'db>, locals: &Locals, ) -> Result<(), MirEvalError>

fn exec_fn_pointer( &mut self, bytes: Interval, destination: Interval, args: &[IntervalAndTy<'db>], locals: &Locals, target_bb: Option<BasicBlockId>, span: MirSpan, ) -> Result<Option<StackFrame>, MirEvalError>

fn exec_closure( &mut self, closure: InternedClosureId, closure_data: Interval, generic_args: GenericArgs<'db>, destination: Interval, args: &[IntervalAndTy<'db>], locals: &Locals, span: MirSpan, ) -> Result<Option<StackFrame>, MirEvalError>

fn exec_fn_def( &mut self, def: CallableDefId, generic_args: GenericArgs<'db>, destination: Interval, args: &[IntervalAndTy<'db>], locals: &Locals, target_bb: Option<BasicBlockId>, span: MirSpan, ) -> Result<Option<StackFrame>, MirEvalError>

fn get_mir_or_dyn_index( &self, def: FunctionId, generic_args: GenericArgs<'db>, locals: &Locals, span: MirSpan, ) -> Result<MirOrDynIndex, MirEvalError>

fn exec_fn_with_args( &mut self, def: FunctionId, args: &[IntervalAndTy<'db>], generic_args: GenericArgs<'db>, locals: &Locals, destination: Interval, target_bb: Option<BasicBlockId>, span: MirSpan, ) -> Result<Option<StackFrame>, MirEvalError>

fn exec_looked_up_function( &mut self, mir_body: Arc<MirBody>, locals: &Locals, def: FunctionId, arg_bytes: impl Iterator<Item = IntervalOrOwned>, span: MirSpan, destination: Interval, target_bb: Option<BasicBlockId>, ) -> Result<Option<StackFrame>, MirEvalError>

fn exec_fn_trait( &mut self, def: FunctionId, args: &[IntervalAndTy<'db>], generic_args: GenericArgs<'db>, locals: &Locals, destination: Interval, target_bb: Option<BasicBlockId>, span: MirSpan, ) -> Result<Option<StackFrame>, MirEvalError>

fn eval_static( &mut self, st: StaticId, locals: &Locals, ) -> Result<Address, MirEvalError>

fn const_eval_discriminant( &self, variant: EnumVariantId, ) -> Result<i128, MirEvalError>

fn drop_place( &mut self, place: &Place, locals: &mut Locals, span: MirSpan, ) -> Result<(), MirEvalError>

fn run_drop_glue_deep( &mut self, ty: Ty<'db>, locals: &Locals, addr: Address, _metadata: &[u8], span: MirSpan, ) -> Result<(), MirEvalError>

fn write_to_stdout(&mut self, interval: Interval) -> Result<(), MirEvalError>

fn write_to_stderr(&mut self, interval: Interval) -> Result<(), MirEvalError>