hir_ty::mir

Module lower

This module generates a polymorphic MIR from a hir body

Modules

as_place 🔒

MIR lowering for places

pattern_matching 🔒

MIR lowering for patterns

Macros

implementation_error 🔒

not_supported 🔒

Structs

DropScope 🔒

DropScopeToken 🔒

A token to ensuring that each drop scope is popped at most once, thanks to the compiler that checks moves.

LoopBlocks 🔒

MirLowerCtx 🔒

Enums

MirLowerError

Functions

cast_kind 🔒

lower_to_mir

mir_body_cycle_result 🔒

mir_body_for_closure_query

mir_body_query

Type Aliases

Result 🔒