hir_ty::next_solver::fulfill::errors::wf

Function unnormalized_obligations

pub(crate) fn unnormalized_obligations<'db>(
    infcx: &InferCtxt<'db>,
    param_env: ParamEnv<'db>,
    term: Term<'db>,
) -> Option<Vec<Obligation<'db, Predicate<'db>>>>

Compute the predicates that are required for a type to be well-formed.

This is only intended to be used in the new solver, since it does not take into account recursion depth or proper error-reporting spans.