hir_ty::next_solver::infer::canonical::canonicalizer

Struct Canonicalizer

struct Canonicalizer<'cx, 'db> {
    infcx: &'cx InferCtxt<'db>,
    tcx: DbInterner<'db>,
    variables: SmallVec<[CanonicalVarKind<'db>; 8]>,
    query_state: &'cx mut OriginalQueryValues<'db>,
    indices: FxHashMap<GenericArg<'db>, BoundVar>,
    sub_root_lookup_table: FxHashMap<TyVid, usize>,
    canonicalize_mode: &'cx dyn CanonicalizeMode,
    needs_canonical_flags: TypeFlags,
    binder_index: DebruijnIndex,
}

Fields
infcx: &'cx InferCtxt<'db>

Set to None to disable the resolution of inference variables.

tcx: DbInterner<'db>variables: SmallVec<[CanonicalVarKind<'db>; 8]>query_state: &'cx mut OriginalQueryValues<'db>indices: FxHashMap<GenericArg<'db>, BoundVar>sub_root_lookup_table: FxHashMap<TyVid, usize>

Maps each sub_unification_table_root_var to the index of the first variable which used it.

This means in case two type variables have the same sub relations root, we set the sub_root of the second variable to the position of the first. Otherwise the sub_root of each type variable is just its own position.

canonicalize_mode: &'cx dyn CanonicalizeModeneeds_canonical_flags: TypeFlagsbinder_index: DebruijnIndex
Implementations

impl<'cx, 'db> Canonicalizer<'cx, 'db>

fn canonicalize<V>( value: V, infcx: &InferCtxt<'db>, tcx: DbInterner<'db>, canonicalize_region_mode: &dyn CanonicalizeMode, query_state: &mut OriginalQueryValues<'db>, ) -> Canonical<'db, V>
where V: TypeFoldable<DbInterner<'db>>,

The main canonicalize method, shared impl of canonicalize_query and canonicalize_response.

fn canonicalize_with_base<U, V>( base: Canonical<'db, U>, value: V, infcx: &InferCtxt<'db>, tcx: DbInterner<'db>, canonicalize_region_mode: &dyn CanonicalizeMode, query_state: &mut OriginalQueryValues<'db>, ) -> Canonical<'db, (U, V)>
where V: TypeFoldable<DbInterner<'db>>,

fn canonical_var( &mut self, info: CanonicalVarKind<'db>, kind: GenericArg<'db>, ) -> BoundVar

Creates a canonical variable replacing kind from the input, or returns an existing variable if kind has already been seen. kind is expected to be an unbound variable (or potentially a free region).

fn get_or_insert_sub_root(&mut self, vid: TyVid) -> BoundVar

fn universe_canonicalized_variables( self, ) -> SmallVec<[CanonicalVarKind<'db>; 8]>

Replaces the universe indexes used in var_values with their index in query_state.universe_map. This minimizes the maximum universe used in the canonicalized value.

fn canonical_var_for_region_in_root_universe( &mut self, r: Region<'db>, ) -> Region<'db>

Shorthand helper that creates a canonical region variable for r (always in the root universe). The reason that we always put these variables into the root universe is because this method is used during query construction: in that case, we are taking all the regions and just putting them into the most generic context we can. This may generate solutions that don’t fit (e.g., that equate some region variable with a placeholder it can’t name) on the caller side, but that’s ok, the caller can figure that out. In the meantime, it maximizes our caching.

(This works because unification never fails – and hence trait selection is never affected – due to a universe mismatch.)

fn canonical_var_for_region( &mut self, info: CanonicalVarKind<'db>, r: Region<'db>, ) -> Region<'db>

Creates a canonical variable (with the given info) representing the region r; return a region referencing it.

fn canonicalize_ty_var( &mut self, info: CanonicalVarKind<'db>, ty_var: Ty<'db>, ) -> Ty<'db>

Given a type variable ty_var of the given kind, first check if ty_var is bound to anything; if so, canonicalize that. Otherwise, create a new canonical variable for ty_var.

fn canonicalize_const_var( &mut self, info: CanonicalVarKind<'db>, const_var: Const<'db>, ) -> Const<'db>

Given a type variable const_var of the given kind, first check if const_var is bound to anything; if so, canonicalize that. Otherwise, create a new canonical variable for const_var.