hir_ty::next_solver::infer::region_constraints

Struct RegionConstraintCollector

pub struct RegionConstraintCollector<'db, 'a> {
    storage: &'a mut RegionConstraintStorage<'db>,
    undo_log: &'a mut InferCtxtUndoLogs<'db>,
}

Fields
storage: &'a mut RegionConstraintStorage<'db>undo_log: &'a mut InferCtxtUndoLogs<'db>
Implementations

impl<'db> RegionConstraintCollector<'db, '_>

pub fn num_region_vars(&self) -> usize

pub fn region_constraint_data(&self) -> &RegionConstraintData<'db>

pub fn take_and_reset_data(&mut self) -> RegionConstraintData<'db>

Takes (and clears) the current set of constraints. Note that the set of variables remains intact, but all relationships between them are reset. This is used during NLL checking to grab the set of constraints that arose from a particular operation.

We don’t want to leak relationships between variables between points because just because (say) r1 == r2 was true at some point P in the graph doesn’t imply that it will be true at some other point Q, in NLL.

Not legal during a snapshot.

pub fn data(&self) -> &RegionConstraintData<'db>

pub(super) fn start_snapshot(&self) -> RegionSnapshot

pub(super) fn rollback_to(&mut self, snapshot: RegionSnapshot)

pub(super) fn new_region_var(&mut self, universe: UniverseIndex) -> RegionVid

fn add_constraint(&mut self, constraint: Constraint<'db>)

pub(super) fn make_eqregion(&mut self, a: Region<'db>, b: Region<'db>)

pub(super) fn make_subregion(&mut self, sub: Region<'db>, sup: Region<'db>)

pub(super) fn lub_regions( &mut self, db: DbInterner<'db>, a: Region<'db>, b: Region<'db>, ) -> Region<'db>

pub(super) fn glb_regions( &mut self, db: DbInterner<'db>, a: Region<'db>, b: Region<'db>, ) -> Region<'db>

pub fn opportunistic_resolve_var( &mut self, cx: DbInterner<'db>, vid: RegionVid, ) -> Region<'db>

Resolves a region var to its value in the unification table, if it exists. Otherwise, it is resolved to the root ReVar in the table.

pub fn probe_value( &mut self, vid: RegionVid, ) -> Result<Region<'db>, UniverseIndex>

fn combine_map( &mut self, t: CombineMapType, ) -> &mut FxHashMap<TwoRegions<'db>, RegionVid>

fn combine_vars( &mut self, cx: DbInterner<'db>, t: CombineMapType, a: Region<'db>, b: Region<'db>, ) -> Region<'db>

pub fn universe(&mut self, region: Region<'db>) -> UniverseIndex

pub fn vars_since_snapshot(&self, value_count: usize) -> Range<RegionVid>

pub fn region_constraints_added_in_snapshot(&self, mark: &Snapshot) -> bool

See InferCtxt::region_constraints_added_in_snapshot.

fn unification_table_mut( &mut self, ) -> UnificationTable<InPlace<RegionVidKey<'db>, &'_ mut UnificationStorage<RegionVidKey<'db>>, &'_ mut InferCtxtUndoLogs<'db>>>