hir_ty::next_solver::infer

Struct InferCtxtInner

pub struct InferCtxtInner<'db> {
    pub(crate) undo_log: InferCtxtUndoLogs<'db>,
    pub(crate) type_variable_storage: TypeVariableStorage<'db>,
    pub(crate) const_unification_storage: UnificationTableStorage<ConstVidKey<'db>>,
    pub(crate) int_unification_storage: UnificationTableStorage<IntVid>,
    pub(crate) float_unification_storage: UnificationTableStorage<FloatVid>,
    pub(crate) region_constraint_storage: Option<RegionConstraintStorage<'db>>,
    pub(crate) region_obligations: Vec<TypeOutlivesConstraint<'db>>,
    region_assumptions: Vec<ArgOutlivesPredicate<'db>>,
    pub(crate) opaque_type_storage: OpaqueTypeStorage<'db>,
}

This type contains all the things within InferCtxt that sit within a RefCell and are involved with taking/rolling back snapshots. Snapshot operations are hot enough that we want only one call to borrow_mut per call to start_snapshot and rollback_to.

Fields
undo_log: InferCtxtUndoLogs<'db>type_variable_storage: TypeVariableStorage<'db>

We instantiate UnificationTable with bounds<Ty> because the types that might instantiate a general type variable have an order, represented by its upper and lower bounds.

const_unification_storage: UnificationTableStorage<ConstVidKey<'db>>

Map from const parameter variable to the kind of const it represents.

int_unification_storage: UnificationTableStorage<IntVid>

Map from integral variable to the kind of integer it represents.

float_unification_storage: UnificationTableStorage<FloatVid>

Map from floating variable to the kind of float it represents.

region_constraint_storage: Option<RegionConstraintStorage<'db>>

Tracks the set of region variables and the constraints between them.

This is initially Some(_) but when resolve_regions_and_report_errors is invoked, this gets set to None โ further attempts to perform unification, etc., may fail if new region constraints wouldโve been added.

region_obligations: Vec<TypeOutlivesConstraint<'db>>

A set of constraints that regionck must validate.

Each constraint has the form T:'a, meaning โsome type T must outlive the lifetime โaโ. These constraints derive from instantiated type parameters. So if you had a struct defined like the following:

struct Foo<T: 'static> { ... }

In some expression let x = Foo { ... }, it will instantiate the type parameter T with a fresh type $0. At the same time, it will record a region obligation of $0: 'static. This will get checked later by regionck. (We canโt generally check these things right away because we have to wait until types are resolved.)

These are stored in a map keyed to the id of the innermost enclosing fn body / static initializer expression. This is because the location where the obligation was incurred can be relevant with respect to which sublifetime assumptions are in place. The reason that we store under the fn-id, and not something more fine-grained, is so that it is easier for regionck to be sure that it has found all the region obligations (otherwise, itโs easy to fail to walk to a particular node-id).

Before running resolve_regions_and_report_errors, the creator of the inference context is expected to invoke InferCtxt::process_registered_region_obligations for each body-id in this map, which will process the obligations within. This is expected to be done โlate enoughโ that all type inference variables have been bound and so forth.

region_assumptions: Vec<ArgOutlivesPredicate<'db>>

The outlives bounds that we assume must hold about placeholders that come from instantiating the binder of coroutine-witnesses. These bounds are deduced from the well-formedness of the witnessโs types, and are necessary because of the way we anonymize the regions in a coroutine, which may cause types to no longer be considered well-formed.

opaque_type_storage: OpaqueTypeStorage<'db>

Caches for opaque type inference.

Implementations

impl<'db> InferCtxtInner<'db>

pub fn rollback_to(&mut self, snapshot: Snapshot)

pub fn commit(&mut self, snapshot: Snapshot)

impl<'db> InferCtxtInner<'db>

fn new() -> InferCtxtInner<'db>

pub fn region_obligations(&self) -> &[TypeOutlivesConstraint<'db>]

fn try_type_variables_probe_ref( &self, vid: TyVid, ) -> Option<&TypeVariableValue<'db>>

fn type_variables(&mut self) -> TypeVariableTable<'_, 'db>

pub(crate) fn opaque_types(&mut self) -> OpaqueTypeTable<'_, 'db>

pub(crate) fn int_unification_table( &mut self, ) -> UnificationTable<InPlace<IntVid, &'_ mut UnificationStorage<IntVid>, &'_ mut InferCtxtUndoLogs<'db>>>

pub(crate) fn float_unification_table( &mut self, ) -> UnificationTable<InPlace<FloatVid, &'_ mut UnificationStorage<FloatVid>, &'_ mut InferCtxtUndoLogs<'db>>>

fn const_unification_table( &mut self, ) -> UnificationTable<InPlace<ConstVidKey<'db>, &'_ mut UnificationStorage<ConstVidKey<'db>>, &'_ mut InferCtxtUndoLogs<'db>>>

pub fn unwrap_region_constraints( &mut self, ) -> RegionConstraintCollector<'db, '_>