hir_ty::next_solver::infer::type_variable

Struct TypeVariableTable

pub(crate) struct TypeVariableTable<'a, 'db> {
    storage: &'a mut TypeVariableStorage<'db>,
    undo_log: &'a mut InferCtxtUndoLogs<'db>,
}

Fields
storage: &'a mut TypeVariableStorage<'db>undo_log: &'a mut InferCtxtUndoLogs<'db>
Implementations

impl<'db> TypeVariableTable<'_, 'db>

pub(crate) fn var_origin(&self, vid: TyVid) -> TypeVariableOrigin

Returns the origin that was given when vid was created.

Note that this function does not return care whether vid has been unified with something else or not.

pub(crate) fn equate(&mut self, a: TyVid, b: TyVid)

Records that a == b, depending on dir.

Precondition: neither a nor b are known.

pub(crate) fn sub_unify(&mut self, a: TyVid, b: TyVid)

Records that a and b are related via subtyping. We don’t track which of the two is the subtype.

Precondition: neither a nor b are known.

pub(crate) fn instantiate(&mut self, vid: TyVid, ty: Ty<'db>)

Instantiates vid with the type ty.

Precondition: vid must not have been previously instantiated.

pub(crate) fn new_var( &mut self, universe: UniverseIndex, origin: TypeVariableOrigin, ) -> TyVid

Creates a new type variable.

diverging: indicates if this is a “diverging” type variable, e.g., one created as the type of a return expression. The code in this module doesn’t care if a variable is diverging, but the main Rust type-checker will sometimes “unify” such variables with the ! or () types.

origin: indicates why the type variable was created. The code in this module doesn’t care, but it can be useful for improving error messages.

pub(crate) fn num_vars(&self) -> usize

Returns the number of type variables created thus far.

pub(crate) fn root_var(&mut self, vid: TyVid) -> TyVid

Returns the “root” variable of vid in the eq_relations equivalence table. All type variables that have been equated will yield the same root variable (per the union-find algorithm), so root_var(a) == root_var(b) implies that a == b (transitively).

pub(crate) fn sub_unification_table_root_var(&mut self, vid: TyVid) -> TyVid

Returns the “root” variable of vid in the sub_unification_table equivalence table. All type variables that have been are related via equality or subtyping will yield the same root variable (per the union-find algorithm), so sub_unification_table_root_var(a) == sub_unification_table_root_var(b) implies that:

exists X. (a <: X || X <: a) && (b <: X || X <: b)

pub(crate) fn probe(&mut self, vid: TyVid) -> TypeVariableValue<'db>

Retrieves the type to which vid has been instantiated, if any.

pub(crate) fn inlined_probe(&mut self, vid: TyVid) -> TypeVariableValue<'db>

An always-inlined variant of probe, for very hot call sites.

fn eq_relations( &mut self, ) -> UnificationTable<InPlace<TyVidEqKey<'db>, &'_ mut UnificationStorage<TyVidEqKey<'db>>, &'_ mut InferCtxtUndoLogs<'db>>>

fn sub_unification_table( &mut self, ) -> UnificationTable<InPlace<TyVidSubKey, &'_ mut UnificationStorage<TyVidSubKey>, &'_ mut InferCtxtUndoLogs<'db>>>

pub(crate) fn vars_since_snapshot( &mut self, value_count: usize, ) -> (Range<TyVid>, Vec<TypeVariableOrigin>)

Returns a range of the type variables created during the snapshot.

pub(crate) fn unresolved_variables(&mut self) -> Vec<TyVid>

Returns indices of all variables that are not yet instantiated.