hir_ty::next_solver

Module inspect

Structs

GoalEvaluation 🔒

When evaluating a goal we also store the original values for the CanonicalVarValues of the canonicalized goal. We use this to map any CanonicalState from the local InferCtxt of the solver query to the InferCtxt of the caller.

InspectCandidate 🔒

InspectConfig 🔒

InspectGoal 🔒

NormalizesToTermHack 🔒

The expected term of a NormalizesTo goal gets replaced with an unconstrained inference variable when computing NormalizesTo goals and we return the nested goals to the caller, who also equates the actual term with the expected.

Probe 🔒

A self-contained computation during trait solving. This either corresponds to a EvalCtxt::probe(_X) call or the root evaluation of a goal.

State 🔒

Some data together with information about how they relate to the input of the canonical query.

Enums

ProbeKind 🔒

What kind of probe we’re in. In case the probe represents a candidate, or the final result of the current goal - via ProbeKind::Root - we also store the [QueryResult].

ProbeStep 🔒

Traits

ProofTreeVisitor 🔒

The public API to interact with proof trees.

Type Aliases

CanonicalState 🔒