hir_ty::next_solver::interner

Struct DbInterner

pub struct DbInterner<'db> {
    pub(crate) db: &'db dyn HirDatabase,
    krate: Option<Crate>,
    lang_items: Option<&'db LangItems>,
}

Fields
db: &'db dyn HirDatabasekrate: Option<Crate>lang_items: Option<&'db LangItems>
Implementations

impl<'db> DbInterner<'db>

pub fn instantiate_bound_regions<T, F>( self, value: Binder<'db, T>, fld_r: F, ) -> (T, FxIndexMap<BoundRegion, Region<'db>>)
where F: FnMut(BoundRegion) -> Region<'db>, T: TypeFoldable<DbInterner<'db>>,

Replaces all regions bound by the given Binder with the results returned by the closure; the closure is expected to return a free region (relative to this binder), and hence the binder is removed in the return type. The closure is invoked once for each unique BoundRegionKind; multiple references to the same BoundRegionKind will reuse the previous result. A map is returned at the end with each bound region and the free region that replaced it.

Panics

This method only replaces late bound regions. Any types or constants bound by value will cause an ICE.

pub fn instantiate_bound_regions_uncached<T, F>( self, value: Binder<'db, T>, replace_regions: F, ) -> T
where F: FnMut(BoundRegion) -> Region<'db>, T: TypeFoldable<DbInterner<'db>>,

pub fn instantiate_bound_regions_with_erased<T>( self, value: Binder<'db, T>, ) -> T
where T: TypeFoldable<DbInterner<'db>>,

Replaces any late-bound regions bound in value with 'erased. Useful in codegen but also method lookup and a few other places where precise region relationships are not required.

impl<'db> DbInterner<'db>

pub(super) fn mk_args(self, args: &[GenericArg<'db>]) -> GenericArgs<'db>

pub(super) fn mk_args_from_iter<I, T>(self, iter: I) -> T::Output
where I: Iterator<Item = T>, T: CollectAndApply<GenericArg<'db>, GenericArgs<'db>>,

impl<'db> DbInterner<'db>

pub fn conjure() -> DbInterner<'db>

pub fn new_no_crate(db: &'db dyn HirDatabase) -> Self

Creates a new interner without an active crate. Good only for interning things, not for trait solving etc.. As a rule of thumb, when you create an InferCtxt, you need to provide the crate (and the block).

Elaboration is a special kind: it needs lang items (for Sized), therefore it needs new_with().

pub fn new_with(db: &'db dyn HirDatabase, krate: Crate) -> DbInterner<'db>

pub fn db(&self) -> &'db dyn HirDatabase

pub fn lang_items(&self) -> &'db LangItems

pub fn default_types<'a>(&self) -> &'a DefaultAny<'db>

impl<'db> DbInterner<'db>

pub fn shift_bound_var_indices<T>(self, bound_vars: usize, value: T) -> T
where T: TypeFoldable<Self>,

pub fn replace_escaping_bound_vars_uncached<T: TypeFoldable<DbInterner<'db>>>( self, value: T, delegate: impl BoundVarReplacerDelegate<'db>, ) -> T

pub fn replace_bound_vars_uncached<T: TypeFoldable<DbInterner<'db>>>( self, value: Binder<'db, T>, delegate: impl BoundVarReplacerDelegate<'db>, ) -> T

pub fn mk_fn_sig<I>( self, inputs: I, output: Ty<'db>, c_variadic: bool, safety: Safety, abi: FnAbi, ) -> FnSig<'db>
where I: IntoIterator<Item = Ty<'db>>,

impl<'db> DbInterner<'db>

pub fn signature_unclosure( self, sig: PolyFnSig<'db>, safety: Safety, ) -> PolyFnSig<'db>

Given a closure signature, returns an equivalent fn signature. Detuples and so forth so e.g., if we have a sig with Fn<(u32, i32)> then you would get a fn(u32, i32). unsafety determines the unsafety of the fn signature. If you pass Safety::Unsafe in the previous example, then you would get an unsafe fn (u32, i32). It cannot convert a closure that requires unsafe.