hir_ty::next_solver::obligation_ctxt

Struct ObligationCtxt

pub struct ObligationCtxt<'a, 'db> {
    pub infcx: &'a InferCtxt<'db>,
    /* private fields */
}

Used if you want to have pleasant experience when dealing with obligations outside of hir or mir typeck.

Fields
infcx: &'a InferCtxt<'db>
Implementations

impl<'a, 'db> ObligationCtxt<'a, 'db>

pub fn new(infcx: &'a InferCtxt<'db>) -> Self

impl<'a, 'db> ObligationCtxt<'a, 'db>

pub fn register_obligation( &mut self, obligation: Obligation<'db, Predicate<'db>>, )

pub fn register_obligations( &mut self, obligations: impl IntoIterator<Item = Obligation<'db, Predicate<'db>>>, )

pub fn register_infer_ok_obligations<T>( &mut self, infer_ok: InferOk<'db, T>, ) -> T

pub fn register_bound( &mut self, cause: ObligationCause, param_env: ParamEnv<'db>, ty: Ty<'db>, def_id: TraitId, )

Requires that ty must implement the trait with def_id in the given environment. This trait must not have any type parameters (except for Self).

pub fn eq<T: ToTrace<'db>>( &mut self, cause: &ObligationCause, param_env: ParamEnv<'db>, expected: T, actual: T, ) -> Result<(), TypeError<'db>>

pub fn sub<T: ToTrace<'db>>( &mut self, cause: &ObligationCause, param_env: ParamEnv<'db>, expected: T, actual: T, ) -> Result<(), TypeError<'db>>

Checks whether expected is a subtype of actual: expected <: actual.

pub fn relate<T: ToTrace<'db>>( &mut self, cause: &ObligationCause, param_env: ParamEnv<'db>, variance: Variance, expected: T, actual: T, ) -> Result<(), TypeError<'db>>

pub fn sup<T: ToTrace<'db>>( &mut self, cause: &ObligationCause, param_env: ParamEnv<'db>, expected: T, actual: T, ) -> Result<(), TypeError<'db>>

Checks whether expected is a supertype of actual: expected :> actual.

pub fn lub<T: ToTrace<'db>>( &mut self, cause: &ObligationCause, param_env: ParamEnv<'db>, expected: T, actual: T, ) -> Result<T, TypeError<'db>>

Computes the least-upper-bound, or mutual supertype, of two values.

pub fn try_evaluate_obligations(&mut self) -> Vec<NextSolverError<'db>>

pub fn evaluate_obligations_error_on_ambiguity( &mut self, ) -> Vec<NextSolverError<'db>>

pub fn into_pending_obligations(self) -> Vec<Obligation<'db, Predicate<'db>>>

Returns the not-yet-processed and stalled obligations from the ObligationCtxt.

Takes ownership of the context as doing operations such as ObligationCtxt::eq afterwards will result in other obligations getting ignored. You can make a new ObligationCtxt if this needs to be done in a loop, for example.

pub fn deeply_normalize<T: TypeFoldable<DbInterner<'db>>>( &self, cause: &ObligationCause, param_env: ParamEnv<'db>, value: T, ) -> Result<T, Vec<NextSolverError<'db>>>

pub fn structurally_normalize_ty( &mut self, cause: &ObligationCause, param_env: ParamEnv<'db>, value: Ty<'db>, ) -> Result<Ty<'db>, Vec<NextSolverError<'db>>>

pub fn structurally_normalize_const( &mut self, cause: &ObligationCause, param_env: ParamEnv<'db>, value: Const<'db>, ) -> Result<Const<'db>, Vec<NextSolverError<'db>>>

pub fn structurally_normalize_term( &mut self, cause: &ObligationCause, param_env: ParamEnv<'db>, value: Term<'db>, ) -> Result<Term<'db>, Vec<NextSolverError<'db>>>