hir_ty::next_solver::ty

Struct Ty

pub struct Ty<'db> {
    pub(super) interned: InternedRef<'db, TyInterned>,
}

Fields
interned: InternedRef<'db, TyInterned>
Implementations

impl Ty<'_>

pub fn store(self) -> StoredTy

impl<'db> Ty<'db>

pub fn new(_interner: DbInterner<'db>, kind: TyKind<'db>) -> Self

pub fn inner(&self) -> &WithCachedTypeInfo<TyKind<'db>>

pub fn new_adt( interner: DbInterner<'db>, adt_id: AdtId, args: GenericArgs<'db>, ) -> Self

pub fn new_param(interner: DbInterner<'db>, id: TypeParamId, index: u32) -> Self

pub fn new_placeholder( interner: DbInterner<'db>, placeholder: PlaceholderTy, ) -> Self

pub fn new_infer(interner: DbInterner<'db>, infer: InferTy) -> Self

pub fn new_int_var(interner: DbInterner<'db>, v: IntVid) -> Self

pub fn new_float_var(interner: DbInterner<'db>, v: FloatVid) -> Self

pub fn new_int(interner: DbInterner<'db>, i: IntTy) -> Self

pub fn new_uint(interner: DbInterner<'db>, ui: UintTy) -> Self

pub fn new_float(interner: DbInterner<'db>, f: FloatTy) -> Self

pub fn new_fresh(interner: DbInterner<'db>, n: u32) -> Self

pub fn new_fresh_int(interner: DbInterner<'db>, n: u32) -> Self

pub fn new_fresh_float(interner: DbInterner<'db>, n: u32) -> Self

pub fn new_empty_tuple(interner: DbInterner<'db>) -> Self

pub fn new_imm_ptr(interner: DbInterner<'db>, ty: Ty<'db>) -> Self

pub fn new_imm_ref( interner: DbInterner<'db>, region: Region<'db>, ty: Ty<'db>, ) -> Self

pub fn new_opaque( interner: DbInterner<'db>, def_id: SolverDefId, args: GenericArgs<'db>, ) -> Self

pub fn primitive_size(self, interner: DbInterner<'db>) -> Size

Returns the Size for primitive types (bool, uint, int, char, float).

pub fn int_size_and_signed(self, interner: DbInterner<'db>) -> (Size, bool)

pub fn walk(self) -> TypeWalker<DbInterner<'db>>

pub fn has_trivial_sizedness( self, tcx: DbInterner<'db>, sizedness: SizedTraitKind, ) -> bool

Fast path helper for testing if a type is Sized or MetaSized.

Returning true means the type is known to implement the sizedness trait. Returning false means nothing – could be sized, might not be.

Note that we could never rely on the fact that a type such as [_] is trivially !Sized because we could be in a type environment with a bound such as [_]: Copy. A function with such a bound obviously never can be called, but that doesn’t mean it shouldn’t typecheck. This is why this method doesn’t return Option<bool>.

pub fn is_trivially_pure_clone_copy(self) -> bool

Fast path helper for primitives which are always Copy and which have a side-effect-free Clone impl.

Returning true means the type is known to be pure and Copy+Clone. Returning false means nothing – could be Copy, might not be.

This is mostly useful for optimizations, as these are the types on which we can replace cloning with dereferencing.

pub fn is_trivially_wf(self, tcx: DbInterner<'db>) -> bool

pub fn is_never(self) -> bool

pub fn is_bool(self) -> bool

pub fn is_char(self) -> bool

pub fn is_scalar(self) -> bool

A scalar type is one that denotes an atomic datum, with no sub-components. (A RawPtr is scalar because it represents a non-managed pointer, so its contents are abstract to rustc.)

pub fn is_infer(self) -> bool

pub fn is_numeric(self) -> bool

pub fn is_str(self) -> bool

pub fn is_unit(self) -> bool

pub fn is_u8(self) -> bool

pub fn is_raw_ptr(self) -> bool

pub fn is_array(self) -> bool

pub fn is_slice(self) -> bool

pub fn is_union(self) -> bool

pub fn boxed_ty(self) -> Option<Ty<'db>>

pub fn as_adt(self) -> Option<(AdtId, GenericArgs<'db>)>

pub fn as_slice(self) -> Option<Ty<'db>>

pub fn ty_vid(self) -> Option<TyVid>

pub fn safe_to_unsafe_fn_ty( interner: DbInterner<'db>, sig: PolyFnSig<'db>, ) -> Ty<'db>

Given a fn type, returns an equivalent unsafe fn type; that is, a fn type that is equivalent in every way for being unsafe.

pub fn builtin_deref(self, explicit: bool) -> Option<Ty<'db>>

Returns the type of *ty.

The parameter explicit indicates if this is an explicit dereference. Some types – notably raw ptrs – can only be dereferenced explicitly.

pub fn references_non_lt_error(self) -> bool

Whether the type contains some non-lifetime, aka. type or const, error type.

pub fn references_only_ty_error(self) -> bool

Whether the type contains a type error (ignoring const and lifetime errors).

pub fn callable_sig( self, interner: DbInterner<'db>, ) -> Option<Binder<'db, FnSig<'db>>>

pub fn as_reference(self) -> Option<(Ty<'db>, Region<'db>, Mutability)>

pub fn as_reference_or_ptr(self) -> Option<(Ty<'db>, Rawness, Mutability)>

pub fn as_tuple(self) -> Option<Tys<'db>>

pub fn dyn_trait(self) -> Option<TraitId>

pub fn strip_references(self) -> Ty<'db>

pub fn strip_reference(self) -> Ty<'db>

pub fn replace_infer_with_error(self, interner: DbInterner<'db>) -> Ty<'db>

Replace infer vars with errors.

This needs to be called for every type that may contain infer vars and is yielded to outside inference, as things other than inference do not expect to see infer vars.

pub fn from_builtin_type(interner: DbInterner<'db>, ty: BuiltinType) -> Ty<'db>

pub fn as_builtin(self) -> Option<BuiltinType>

pub fn impl_trait_bounds( self, db: &'db dyn HirDatabase, ) -> Option<Vec<Clause<'db>>>

pub fn equals_ctor(self, other: Ty<'db>) -> bool

FIXME: Get rid of this, it’s not a good abstraction