hir_ty
Struct BoundVar

pub struct BoundVar {
    pub debruijn: DebruijnIndex,
    pub index: usize,
}

Identifies a particular bound variable within a binder. Variables are identified by the combination of a DebruijnIndex, which identifies the binder, and an index within that binder.

Consider this case:

forall<'a, 'b> { forall<'c, 'd> { ... } }

Within the ... term:

the variable 'a have a debruijn index of 1 and index 0

the variable 'b have a debruijn index of 1 and index 1

the variable 'c have a debruijn index of 0 and index 0

the variable 'd have a debruijn index of 0 and index 1

The variables 'a and 'b both have debruijn index of 1 because, counting out, they are the 2nd binder enclosing .... The indices identify the location within that binder.

The variables 'c and 'd both have debruijn index of 0 because they appear in the innermost binder enclosing the .... The indices identify the location within that binder.

Fields
debruijn: DebruijnIndex

Debruijn index, which identifies the binder.

index: usize

Index within the binder.

Implementations

impl BoundVar

pub fn new(debruijn: DebruijnIndex, index: usize) -> BoundVar

Creates a new bound variable.

pub fn to_ty<I>(self, interner: I) -> Ty<I>
where I: Interner,

Casts the bound variable to a type.

pub fn to_lifetime<I>(self, interner: I) -> Lifetime<I>
where I: Interner,

Wrap the bound variable in a lifetime.

pub fn to_const<I>(self, interner: I, ty: Ty<I>) -> Const<I>
where I: Interner,

Wraps the bound variable in a constant.

pub fn bound_within(self, outer_binder: DebruijnIndex) -> bool

True if this variable is bound within the amount innermost binders.

pub fn shifted_in(self) -> BoundVar

Adjusts the debruijn index (see DebruijnIndex::shifted_in).

pub fn shifted_in_from(self, outer_binder: DebruijnIndex) -> BoundVar

Adjusts the debruijn index (see DebruijnIndex::shifted_in).

pub fn shifted_out(self) -> Option<BoundVar>

Adjusts the debruijn index (see DebruijnIndex::shifted_in).

pub fn shifted_out_to(self, outer_binder: DebruijnIndex) -> Option<BoundVar>

Adjusts the debruijn index (see DebruijnIndex::shifted_in).

pub fn index_if_innermost(self) -> Option<usize>

Return the index of the bound variable, but only if it is bound at the innermost binder. Otherwise, returns None.

pub fn index_if_bound_at(self, debruijn: DebruijnIndex) -> Option<usize>

Return the index of the bound variable, but only if it is bound at the innermost binder. Otherwise, returns None.