hir_ty

Struct InferenceResult

pub struct InferenceResult {

    method_resolutions: FxHashMap<ExprId, (FunctionId, StoredGenericArgs)>,
    field_resolutions: FxHashMap<ExprId, Either<FieldId, TupleFieldId>>,
    variant_resolutions: FxHashMap<ExprOrPatId, VariantId>,
    assoc_resolutions: FxHashMap<ExprOrPatId, (CandidateId, StoredGenericArgs)>,
    tuple_field_access_types: ThinVec<StoredTys>,
    pub(crate) type_of_expr: ArenaMap<ExprId, StoredTy>,
    pub(crate) type_of_pat: ArenaMap<PatId, StoredTy>,
    pub(crate) type_of_binding: ArenaMap<BindingId, StoredTy>,
    pub(crate) type_of_type_placeholder: FxHashMap<TypeRefId, StoredTy>,
    pub(crate) type_of_opaque: FxHashMap<InternedOpaqueTyId, StoredTy>,
    pub(crate) type_mismatches: Option<Box<FxHashMap<ExprOrPatId, TypeMismatch>>>,
    pub(crate) has_errors: bool,
    diagnostics: ThinVec<InferenceDiagnostic>,
    error_ty: StoredTy,
    pub(crate) expr_adjustments: FxHashMap<ExprId, Box<[Adjustment]>>,
    pub(crate) pat_adjustments: FxHashMap<PatId, Vec<StoredTy>>,
    pub(crate) binding_modes: ArenaMap<PatId, BindingMode>,
    pub(crate) closure_info: FxHashMap<InternedClosureId, (Vec<CapturedItem>, FnTrait)>,
    pub mutated_bindings_in_closure: FxHashSet<BindingId>,
    pub(crate) coercion_casts: FxHashSet<ExprId>,

}

The result of type inference: A mapping from expressions and patterns to types.

When you add a field that stores types (including Substitution and the like), don’t forget resolve_completely()’ing them in InferenceContext::resolve_all(). Inference variables must not appear in the final inference result.

Fields
method_resolutions: FxHashMap<ExprId, (FunctionId, StoredGenericArgs)>

For each method call expr, records the function it resolves to.

field_resolutions: FxHashMap<ExprId, Either<FieldId, TupleFieldId>>

For each field access expr, records the field it resolves to.

variant_resolutions: FxHashMap<ExprOrPatId, VariantId>

For each struct literal or pattern, records the variant it resolves to.

assoc_resolutions: FxHashMap<ExprOrPatId, (CandidateId, StoredGenericArgs)>

For each associated item record what it resolves to

tuple_field_access_types: ThinVec<StoredTys>

Whenever a tuple field expression access a tuple field, we allocate a tuple id in InferenceContext and store the tuples substitution there. This map is the reverse of that which allows us to resolve a [TupleFieldId]s type.

type_of_expr: ArenaMap<ExprId, StoredTy>type_of_pat: ArenaMap<PatId, StoredTy>

For each pattern record the type it resolves to.

Note: When a pattern type is resolved it may still contain unresolved or missing subpatterns or subpatterns of mismatched types.

type_of_binding: ArenaMap<BindingId, StoredTy>type_of_type_placeholder: FxHashMap<TypeRefId, StoredTy>type_of_opaque: FxHashMap<InternedOpaqueTyId, StoredTy>type_mismatches: Option<Box<FxHashMap<ExprOrPatId, TypeMismatch>>>has_errors: bool

Whether there are any type-mismatching errors in the result.

diagnostics: ThinVec<InferenceDiagnostic>

During inference this field is empty and InferenceContext::diagnostics is filled instead.

error_ty: StoredTy

Interned Error type to return references to.

expr_adjustments: FxHashMap<ExprId, Box<[Adjustment]>>pat_adjustments: FxHashMap<PatId, Vec<StoredTy>>

Stores the types which were implicitly dereferenced in pattern binding modes.

binding_modes: ArenaMap<PatId, BindingMode>

Stores the binding mode (ref in let ref x = 2) of bindings.

This one is tied to the PatId instead of BindingId, because in some rare cases, a binding in an or pattern can have multiple binding modes. For example:

fn foo(mut slice: &[u32]) -> usize {
    slice = match slice {
        [0, rest @ ..] | rest => rest,
    };
    0
}

the first rest has implicit ref binding mode, but the second rest binding mode is move.

closure_info: FxHashMap<InternedClosureId, (Vec<CapturedItem>, FnTrait)>mutated_bindings_in_closure: FxHashSet<BindingId>coercion_casts: FxHashSet<ExprId>
Implementations

impl InferenceResult

pub fn for_body(db: &dyn HirDatabase, def: DefWithBodyId) -> &InferenceResult

impl InferenceResult

fn new(error_ty: Ty<'_>) -> Self

pub fn method_resolution<'db>( &self, expr: ExprId, ) -> Option<(FunctionId, GenericArgs<'db>)>

pub fn field_resolution( &self, expr: ExprId, ) -> Option<Either<FieldId, TupleFieldId>>

pub fn variant_resolution_for_expr(&self, id: ExprId) -> Option<VariantId>

pub fn variant_resolution_for_pat(&self, id: PatId) -> Option<VariantId>

pub fn variant_resolution_for_expr_or_pat( &self, id: ExprOrPatId, ) -> Option<VariantId>

pub fn assoc_resolutions_for_expr<'db>( &self, id: ExprId, ) -> Option<(CandidateId, GenericArgs<'db>)>

pub fn assoc_resolutions_for_pat<'db>( &self, id: PatId, ) -> Option<(CandidateId, GenericArgs<'db>)>

pub fn assoc_resolutions_for_expr_or_pat<'db>( &self, id: ExprOrPatId, ) -> Option<(CandidateId, GenericArgs<'db>)>

pub fn type_mismatch_for_expr(&self, expr: ExprId) -> Option<&TypeMismatch>

pub fn type_mismatch_for_pat(&self, pat: PatId) -> Option<&TypeMismatch>

pub fn type_mismatches( &self, ) -> impl Iterator<Item = (ExprOrPatId, &TypeMismatch)>

pub fn expr_type_mismatches( &self, ) -> impl Iterator<Item = (ExprId, &TypeMismatch)>

pub fn placeholder_types<'db>( &self, ) -> impl Iterator<Item = (TypeRefId, Ty<'db>)>

pub fn type_of_type_placeholder<'db>( &self, type_ref: TypeRefId, ) -> Option<Ty<'db>>

pub fn closure_info( &self, closure: InternedClosureId, ) -> &(Vec<CapturedItem>, FnTrait)

pub fn type_of_expr_or_pat<'db>(&self, id: ExprOrPatId) -> Option<Ty<'db>>

pub fn type_of_expr_with_adjust<'db>(&self, id: ExprId) -> Option<Ty<'db>>

pub fn type_of_pat_with_adjust<'db>(&self, id: PatId) -> Option<Ty<'db>>

pub fn is_erroneous(&self) -> bool

pub fn diagnostics(&self) -> &[InferenceDiagnostic]

pub fn tuple_field_access_type<'db>(&self, id: TupleId) -> Tys<'db>

pub fn pat_adjustment(&self, id: PatId) -> Option<&[StoredTy]>

pub fn expr_adjustment(&self, id: ExprId) -> Option<&[Adjustment]>

pub fn binding_mode(&self, id: PatId) -> Option<BindingMode>

pub fn expression_types<'db>(&self) -> impl Iterator<Item = (ExprId, Ty<'db>)>

pub fn pattern_types<'db>(&self) -> impl Iterator<Item = (PatId, Ty<'db>)>

pub fn binding_types<'db>(&self) -> impl Iterator<Item = (BindingId, Ty<'db>)>

pub fn return_position_impl_trait_types<'db>( &'db self, db: &'db dyn HirDatabase, ) -> impl Iterator<Item = (ImplTraitIdx, Ty<'db>)>

pub fn expr_ty<'db>(&self, id: ExprId) -> Ty<'db>

pub fn pat_ty<'db>(&self, id: PatId) -> Ty<'db>

pub fn expr_or_pat_ty<'db>(&self, id: ExprOrPatId) -> Ty<'db>

pub fn binding_ty<'db>(&self, id: BindingId) -> Ty<'db>