Struct hir_ty::Ty

pub struct Ty(_);

Implementations

impl Ty

pub fn def_crates(
&self,
db: &dyn HirDatabase,
cur_crate: CrateId
) -> Option<ArrayVec<CrateId, 2>>

impl Ty

pub fn kind(&self, _interner: &Interner) -> &TyKind

pub fn interned_mut(&mut self) -> &mut TyKind

pub fn into_inner(self) -> TyKind

impl Ty

pub fn unit() -> Self

pub fn adt_ty(adt: AdtId, substs: Substitution) -> Ty

pub fn fn_ptr(sig: CallableSig) -> Self

pub fn builtin(builtin: BuiltinType) -> Self

pub fn as_reference(&self) -> Option<(&Ty, Mutability)>

pub fn as_reference_or_ptr(&self) -> Option<(&Ty, Rawness, Mutability)>

pub fn strip_references(&self) -> &Ty

pub fn as_adt(&self) -> Option<(AdtId, &Substitution)>

pub fn as_tuple(&self) -> Option<&Substitution>

pub fn as_generic_def(&self, db: &dyn HirDatabase) -> Option<GenericDefId>

pub fn is_never(&self) -> bool

pub fn is_unknown(&self) -> bool

pub fn equals_ctor(&self, other: &Ty) -> bool

pub fn dyn_trait(&self) -> Option<TraitId>

If this is a dyn Trait, returns that trait.

pub fn callable_def(&self, db: &dyn HirDatabase) -> Option<CallableDefId>

pub fn as_fn_def(&self, db: &dyn HirDatabase) -> Option<FunctionId>

pub fn callable_sig(&self, db: &dyn HirDatabase) -> Option<CallableSig>

pub fn substs(&self) -> Option<&Substitution>

Returns the type parameters of this type if it has some (i.e. is an ADT or function); so if self is Option<u32>, this returns the u32.

pub fn impl_trait_bounds(
&self,
db: &dyn HirDatabase
) -> Option<Vec<QuantifiedWhereClause>>

pub fn associated_type_parent_trait(
&self,
db: &dyn HirDatabase
) -> Option<TraitId>