hir_ty

Module utils

Helper functions for working with def, which don’t need to be a separate query, but can’t be computed directly from *Data (ie, which need a db).

Enums

TargetFeatureIsSafeInTarget

Unsafety

Functions

all_super_traits

Returns the whole super trait hierarchy (including the trait itself).

detect_variant_from_bytes 🔒

direct_super_traits

Returns an iterator over the direct super traits (including the trait itself).

fn_traits 🔒

is_fn_unsafe_to_call

target_feature_is_safe_in_target

unsafe_update_eq 🔒 ⚠

SAFETY: old_pointer must be valid for unique writes