Module hir_ty::db

The home of HirDatabase, which is the Salsa database containing all the type inference-related queries.

Structs

AdtDatumQuery

AdtVarianceQuery

AssociatedTyDataQuery

AssociatedTyValueQuery

BorrowckQuery

CallableItemSignatureQuery

ConstEvalDiscriminantQuery

ConstEvalQuery

ConstEvalStaticQuery

ConstParamTyQuery

DynCompatibilityOfTraitQuery

FieldTypesQuery

FnDefDatumQuery

FnDefVarianceQuery

GenericDefaultsQuery

GenericPredicatesForParamQuery

GenericPredicatesQuery

GenericPredicatesWithoutParentQuery

HirDatabaseGroupStorage__

HirDatabaseStorage

Representative struct for the query group.

ImplDatumQuery

ImplSelfTyQuery

ImplTraitQuery

IncoherentInherentImplCratesQuery

InferQuery

InherentImplsInBlockQuery

InherentImplsInCrateQuery

InternCallableDefLookupQuery

InternCallableDefQuery

InternClosureLookupQuery

InternClosureQuery

InternCoroutineLookupQuery

InternCoroutineQuery

InternImplTraitIdLookupQuery

InternImplTraitIdQuery

InternLifetimeParamIdLookupQuery

InternLifetimeParamIdQuery

InternTypeOrConstParamIdLookupQuery

InternTypeOrConstParamIdQuery

InternedCallableDefId

This exists just for Chalk, because Chalk just has a single FnDefId where we have different IDs for struct and enum variant constructors.

InternedClosure

InternedClosureId

InternedConstParamId

InternedCoroutine

InternedCoroutineId

InternedLifetimeParamId

InternedOpaqueTyId

InternedTypeOrConstParamId

LayoutOfAdtQuery

LayoutOfTyQuery

LookupImplMethodQuery

MirBodyForClosureQuery

MirBodyQuery

MonomorphizedMirBodyForClosureQuery

MonomorphizedMirBodyQuery

ProgramClausesForChalkEnvQuery

ReturnTypeImplTraitsQuery

TargetDataLayoutQuery

TraitDatumQuery

TraitEnvironmentQuery

TraitImplsInBlockQuery

TraitImplsInCrateQuery

TraitImplsInDepsQuery

TraitSolveQuery

TyQuery

TypeAliasImplTraitsQuery

ValueTyQuery

Traits

HirDatabase