hir_ty
Module db

The home of HirDatabase, which is the Salsa database containing all the type inference-related queries.

Structs

AdtDatumQuery

AdtVarianceQuery

AssociatedTyDataQuery

AssociatedTyValueQuery

BorrowckQuery

CallableItemSignatureQuery

ConstEvalDiscriminantQuery

ConstEvalQuery

ConstEvalStaticQuery

ConstParamTyQuery

ConstParamTyWithDiagnosticsQuery

DynCompatibilityOfTraitQuery

FieldTypesQuery

FieldTypesWithDiagnosticsQuery

FnDefDatumQuery

FnDefVarianceQuery

GenericDefaultsQuery

GenericDefaultsWithDiagnosticsQuery

GenericPredicatesForParamQuery

GenericPredicatesQuery

GenericPredicatesWithoutParentQuery

GenericPredicatesWithoutParentWithDiagnosticsQuery

HirDatabaseGroupStorage__

HirDatabaseStorage

Representative struct for the query group.

ImplDatumQuery

ImplSelfTyQuery

ImplSelfTyWithDiagnosticsQuery

ImplTraitQuery

ImplTraitWithDiagnosticsQuery

IncoherentInherentImplCratesQuery

InferQuery

InherentImplsInBlockQuery

InherentImplsInCrateQuery

InternCallableDefLookupQuery

InternCallableDefQuery

InternClosureLookupQuery

InternClosureQuery

InternCoroutineLookupQuery

InternCoroutineQuery

InternImplTraitIdLookupQuery

InternImplTraitIdQuery

InternLifetimeParamIdLookupQuery

InternLifetimeParamIdQuery

InternTypeOrConstParamIdLookupQuery

InternTypeOrConstParamIdQuery

InternedCallableDefId

This exists just for Chalk, because Chalk just has a single FnDefId where we have different IDs for struct and enum variant constructors.

InternedClosure

InternedClosureId

InternedConstParamId

InternedCoroutine

InternedCoroutineId

InternedLifetimeParamId

InternedOpaqueTyId

InternedTypeOrConstParamId

LayoutOfAdtQuery

LayoutOfTyQuery

LookupImplMethodQuery

MirBodyForClosureQuery

MirBodyQuery

MonomorphizedMirBodyForClosureQuery

MonomorphizedMirBodyQuery

ProgramClausesForChalkEnvQuery

ReturnTypeImplTraitsQuery

TargetDataLayoutQuery

TraitDatumQuery

TraitEnvironmentQuery

TraitImplsInBlockQuery

TraitImplsInCrateQuery

TraitImplsInDepsQuery

TraitSolveQuery

TyQuery

TypeAliasImplTraitsQuery

TypeForTypeAliasWithDiagnosticsQuery

ValueTyQuery

VariancesOfQuery

Traits

HirDatabase