Module hir_ty::db

The home of HirDatabase, which is the Salsa database containing all the type inference-related queries.

Structs

AdtVarianceQuery

AssociatedTyDataQuery

AssociatedTyValueQuery

CallableItemSignatureQuery

ConstEvalQuery

ConstParamTyQuery

FieldTypesQuery

FnDefDatumQuery

FnDefVarianceQuery

GenericDefaultsQuery

GenericPredicatesForParamQuery

GenericPredicatesQuery

HirDatabaseGroupStorage__

HirDatabaseStorage

Representative struct for the query group.

ImplDatumQuery

ImplSelfTyQuery

ImplTraitQuery

InferQueryQuery

InherentImplCratesQuery

InherentImplsInBlockQuery

InherentImplsInCrateQuery

InternCallableDefLookupQuery

InternCallableDefQuery

InternClosureLookupQuery

InternClosureQuery

InternImplTraitIdLookupQuery

InternImplTraitIdQuery

InternLifetimeParamIdLookupQuery

InternLifetimeParamIdQuery

InternTypeOrConstParamIdLookupQuery

InternTypeOrConstParamIdQuery

InternedCallableDefId

This exists just for Chalk, because Chalk just has a single FnDefId where we have different IDs for struct and enum variant constructors.

InternedClosureId

InternedConstParamId

InternedLifetimeParamId

InternedOpaqueTyId

InternedTypeOrConstParamId

ProgramClausesForChalkEnvQuery

ReturnTypeImplTraitsQuery

StructDatumQuery

TraitDatumQuery

TraitEnvironmentQuery

TraitImplsInBlockQuery

TraitImplsInCrateQuery

TraitImplsInDepsQuery

TraitSolveQueryQuery

TyQuery

ValueTyQuery

Traits

HirDatabase