Module hir_ty::db

The home of HirDatabase, which is the Salsa database containing all the type inference-related queries.

Structs

AdtVarianceQuery

AssociatedTyDataQuery

AssociatedTyValueQuery

BorrowckQuery

CallableItemSignatureQuery

ConstEvalDiscriminantQuery

ConstEvalQuery

ConstParamTyQuery

FieldTypesQuery

FnDefDatumQuery

FnDefVarianceQuery

GenericDefaultsQuery

GenericPredicatesForParamQuery

GenericPredicatesQuery

HirDatabaseGroupStorage__

HirDatabaseStorage

Representative struct for the query group.

ImplDatumQuery

ImplSelfTyQuery

ImplTraitQuery

IncoherentInherentImplCratesQuery

InferQueryQuery

InherentImplsInBlockQuery

InherentImplsInCrateQuery

InternCallableDefLookupQuery

InternCallableDefQuery

InternClosureLookupQuery

InternClosureQuery

InternGeneratorLookupQuery

InternGeneratorQuery

InternImplTraitIdLookupQuery

InternImplTraitIdQuery

InternLifetimeParamIdLookupQuery

InternLifetimeParamIdQuery

InternTypeOrConstParamIdLookupQuery

InternTypeOrConstParamIdQuery

InternedCallableDefId

This exists just for Chalk, because Chalk just has a single FnDefId where we have different IDs for struct and enum variant constructors.

InternedClosureId

InternedConstParamId

InternedGeneratorId

InternedLifetimeParamId

InternedOpaqueTyId

InternedTypeOrConstParamId

LayoutOfAdtQuery

MirBodyForClosureQuery

MirBodyQuery

ProgramClausesForChalkEnvQuery

ReturnTypeImplTraitsQuery

StructDatumQuery

TargetDataLayoutQuery

TraitDatumQuery

TraitEnvironmentQuery

TraitImplsInBlockQuery

TraitImplsInCrateQuery

TraitImplsInDepsQuery

TraitSolveQueryQuery

TyQuery

ValueTyQuery

Traits

HirDatabase