Module hir_ty::db

The home of HirDatabase, which is the Salsa database containing all the type inference-related queries.

Structs

AdtVarianceQuery

AssociatedTyDataQuery

AssociatedTyValueQuery

CallableItemSignatureQuery

ConstParamTyQuery

FieldTypesQuery

FnDefDatumQuery

FnDefVarianceQuery

GenericDefaultsQuery

GenericPredicatesForParamQuery

GenericPredicatesQuery

HirDatabaseGroupStorage__

HirDatabaseStorage
Representative struct for the query group.

ImplDatumQuery

ImplSelfTyQuery

ImplTraitQuery

InferQueryQuery

InherentImplsInCrateQuery

InternCallableDefLookupQuery

InternCallableDefQuery

InternClosureLookupQuery

InternClosureQuery

InternConstParamIdLookupQuery

InternConstParamIdQuery

InternImplTraitIdLookupQuery

InternImplTraitIdQuery

InternLifetimeParamIdLookupQuery

InternLifetimeParamIdQuery

InternTypeParamIdLookupQuery

InternTypeParamIdQuery

InternedCallableDefId
This exists just for Chalk, because Chalk just has a single FnDefId where we have different IDs for struct and enum variant constructors.

InternedClosureId

InternedConstParamId

InternedLifetimeParamId

InternedOpaqueTyId

InternedTypeParamId

ProgramClausesForChalkEnvQuery

ReturnTypeImplTraitsQuery

StructDatumQuery

TraitDatumQuery

TraitEnvironmentQuery

TraitImplsInCrateQuery

TraitImplsInDepsQuery

TraitSolveQueryQuery

TyQuery

ValueTyQuery

Traits

HirDatabase