hir_ty::layout

Type Alias Layout

pub type Layout = LayoutData<RustcFieldIdx, RustcEnumVariantIdx>;

Aliased Type

pub struct Layout {
    pub fields: FieldsShape<RustcFieldIdx>,
    pub variants: Variants<RustcFieldIdx, RustcEnumVariantIdx>,
    pub backend_repr: BackendRepr,
    pub largest_niche: Option<Niche>,
    pub uninhabited: bool,
    pub align: AbiAlign,
    pub size: Size,
    pub max_repr_align: Option<Align>,
    pub unadjusted_abi_align: Align,
    pub randomization_seed: Hash64,
}

Fields
fields: FieldsShape<RustcFieldIdx>

Says where the fields are located within the layout.

variants: Variants<RustcFieldIdx, RustcEnumVariantIdx>

Encodes information about multi-variant layouts. Even with Multiple variants, a layout still has its own fields! Those are then shared between all variants. One of them will be the discriminant, but e.g. coroutines can have more.

To access all fields of this layout, both fields and the fields of the active variant must be taken into account.

backend_repr: BackendRepr

The backend_repr defines how this data will be represented to the codegen backend, and encodes value restrictions via valid_range.

Note that this is entirely orthogonal to the recursive structure defined by variants and fields; for example, ManuallyDrop<Result<isize, isize>> has IrForm::ScalarPair! So, even with non-Memory backend_repr, fields and variants have to be taken into account to find all fields of this layout.

largest_niche: Option<Niche>

The leaf scalar with the largest number of invalid values (i.e. outside of its valid_range), if it exists.

uninhabited: bool

Is this type known to be uninhabted?

This is separate from BackendRepr because uninhabited return types can affect ABI, especially in the case of by-pointer struct returns, which allocate stack even when unused.

align: AbiAlignsize: Sizemax_repr_align: Option<Align>

The largest alignment explicitly requested with repr(align) on this type or any field. Only used on i686-windows, where the argument passing ABI is different when alignment is requested, even if the requested alignment is equal to the natural alignment.

unadjusted_abi_align: Align

The alignment the type would have, ignoring any repr(align) but including repr(packed). Only used on aarch64-linux, where the argument passing ABI ignores the requested alignment in some cases.

randomization_seed: Hash64

The randomization seed based on this type’s own repr and its fields.

Since randomization is toggled on a per-crate basis even crates that do not have randomization enabled should still calculate a seed so that downstream uses can use it to distinguish different types.

For every T and U for which we do not guarantee that a repr(Rust) Foo<T> can be coerced or transmuted to Foo<U> we aim to create probalistically distinct seeds so that Foo can choose to reorder its fields based on that information. The current implementation is a conservative approximation of this goal.