hir_ty

Module method_resolution

This module is concerned with finding methods that a given type provides. For details about how this works in rustc, see the method lookup page in the rustc guide and the corresponding code mostly in rustc_hir_analysis/check/method/probe.rs.

Structs

InherentImpls

Inherent impls defined in some crate.

ReceiverAdjustments

TraitImpls

Trait impls defined or available in some crate.

Enums

LookupMode

Whether we’re looking up a dotted method call (like v.len()) or a path (like Vec::new).

TyFingerprint

This is used as a key for indexing impls.

VisibleFromModule

Traits

MethodCandidateCallback

Functions

check_orphan_rules

Checks whether the impl satisfies the orphan rules.

def_crates

implements_trait

implements_trait_unique

is_dyn_method

Checks if the self parameter of Trait method is the dyn Trait and we should call the method using the vtable.

iterate_method_candidates_dyn

iterate_path_candidates

lookup_impl_const