hir_ty

Module method_resolution

This module is concerned with finding methods that a given type provides. For details about how this works in rustc, see the method lookup page in the rustc guide and the corresponding code mostly in rustc_hir_analysis/check/method/probe.rs.

Structs

Candidate

CandidateStep

CandidateWithPrivate

InherentImpls

MethodResolutionContext

MethodResolutionUnstableFeatures

Pick

TraitImpls

Enums

CandidateId

CandidateKind

CandidateSource

MethodError

Mode

PickKind

Functions

incoherent_inherent_impls

is_dyn_method

Checks if the self parameter of Trait method is the dyn Trait and we should call the method using the vtable.

lookup_impl_const

simplified_type_module