Module hir_ty::method_resolution

This module is concerned with finding methods that a given type provides. For details about how this works in rustc, see the method lookup page in the rustc guide and the corresponding code mostly in librustc_typeck/check/method/probe.rs.

Structs

InherentImpls
Inherent impls defined in some crate.

TraitImpls
Trait impls defined or available in some crate.

Enums

LookupMode
Whether we’re looking up a dotted method call (like v.len()) or a path (like Vec::new).

TyFingerprint
This is used as a key for indexing impls.

Functions

def_crates

implements_trait

implements_trait_unique

iterate_method_candidates

resolve_indexing_op
Returns the self type for the index trait call.