Struct hir_ty::TyBuilder

pub struct TyBuilder<D> { /* private fields */ }

This is a builder for Ty or anything that needs a Substitution.

Implementations

impl<D> TyBuilder<D>

pub fn push(self, arg: impl CastTo<GenericArg>) -> Self

pub fn remaining(&self) -> usize

pub fn fill_with_bound_vars( self, debruijn: DebruijnIndex, starting_from: usize ) -> Self

pub fn fill_with_unknown(self) -> Self

pub fn fill(self, filler: impl FnMut(&ParamKind) -> GenericArg) -> Self

impl TyBuilder<()>

pub fn unit() -> Ty

pub fn discr_ty() -> Ty

pub fn bool() -> Ty

pub fn usize() -> Ty

pub fn fn_ptr(sig: CallableSig) -> Ty

pub fn builtin(builtin: BuiltinType) -> Ty

pub fn slice(argument: Ty) -> Ty

pub fn placeholder_subst( db: &dyn HirDatabase, def: impl Into<GenericDefId> ) -> Substitution

pub fn subst_for_def( db: &dyn HirDatabase, def: impl Into<GenericDefId>, parent_subst: Option<Substitution> ) -> TyBuilder<()>

pub fn subst_for_generator( db: &dyn HirDatabase, parent: DefWithBodyId ) -> TyBuilder<()>

Creates a TyBuilder to build Substitution for a generator defined in parent.

A generator’s substitution consists of:

resume type of generator

yield type of generator (Generator::Yield)

return type of generator (Generator::Return)

generic parameters in scope on parent in this order.

This method prepopulates the builder with placeholder substitution of parent, so you should only push exactly 3 GenericArgs before building.

pub fn build(self) -> Substitution

impl TyBuilder<AdtId>

pub fn adt(db: &dyn HirDatabase, def: AdtId) -> TyBuilder<AdtId>

pub fn fill_with_defaults( self, db: &dyn HirDatabase, fallback: impl FnMut() -> Ty ) -> Self

pub fn build(self) -> Ty

impl TyBuilder<Tuple>

pub fn tuple(size: usize) -> TyBuilder<Tuple>

pub fn build(self) -> Ty

pub fn tuple_with<I>(elements: I) -> Tywhere I: IntoIterator<Item = Ty>, <I as IntoIterator>::IntoIter: ExactSizeIterator,

impl TyBuilder<TraitId>

pub fn trait_ref(db: &dyn HirDatabase, def: TraitId) -> TyBuilder<TraitId>

pub fn build(self) -> TraitRef

impl TyBuilder<TypeAliasId>

pub fn assoc_type_projection( db: &dyn HirDatabase, def: TypeAliasId, parent_subst: Option<Substitution> ) -> TyBuilder<TypeAliasId>

pub fn build(self) -> ProjectionTy

impl<T: HasInterner<Interner = Interner> + TypeFoldable<Interner>> TyBuilder<Binders<T>>

pub fn build(self) -> T

impl TyBuilder<Binders<Ty>>

pub fn def_ty( db: &dyn HirDatabase, def: TyDefId, parent_subst: Option<Substitution> ) -> TyBuilder<Binders<Ty>>

pub fn impl_self_ty(db: &dyn HirDatabase, def: ImplId) -> TyBuilder<Binders<Ty>>

pub fn value_ty( db: &dyn HirDatabase, def: ValueTyDefId, parent_subst: Option<Substitution> ) -> TyBuilder<Binders<Ty>>