hir_ty
Struct TyLoweringContext

pub struct TyLoweringContext<'a> {
    pub db: &'a dyn HirDatabase,
    pub type_param_mode: ParamLoweringMode,
    /* private fields */
}

Fields
db: &'a dyn HirDatabasetype_param_mode: ParamLoweringMode

Note: Conceptually, it’s thinkable that we could be in a location where some type params should be represented as placeholders, and others should be converted to variables. I think in practice, this isn’t possible currently, so this should be fine for now.

Implementations

impl<'a> TyLoweringContext<'a>

pub fn new( db: &'a dyn HirDatabase, resolver: &'a Resolver, types_map: &'a TypesMap, owner: TypeOwnerId, ) -> Self

pub fn new_maybe_unowned( db: &'a dyn HirDatabase, resolver: &'a Resolver, types_map: &'a TypesMap, types_source_map: Option<&'a TypesSourceMap>, owner: Option<TypeOwnerId>, ) -> Self

pub fn with_debruijn<T>( &mut self, debruijn: DebruijnIndex, f: impl FnOnce(&mut TyLoweringContext<'_>) -> T, ) -> T

pub fn with_shifted_in<T>( &mut self, debruijn: DebruijnIndex, f: impl FnOnce(&mut TyLoweringContext<'_>) -> T, ) -> T

pub fn with_impl_trait_mode( self, impl_trait_mode: ImplTraitLoweringMode, ) -> Self

pub fn with_type_param_mode(self, type_param_mode: ParamLoweringMode) -> Self

pub fn impl_trait_mode( &mut self, impl_trait_mode: ImplTraitLoweringMode, ) -> &mut Self

pub fn type_param_mode( &mut self, type_param_mode: ParamLoweringMode, ) -> &mut Self

pub fn push_diagnostic( &mut self, type_ref: TypeRefId, kind: TyLoweringDiagnosticKind, )

impl<'a> TyLoweringContext<'a>

pub fn lower_ty(&mut self, type_ref: TypeRefId) -> Ty

pub fn lower_const(&mut self, const_ref: &ConstRef, const_type: Ty) -> Const

pub fn lower_ty_ext(&mut self, type_ref_id: TypeRefId) -> (Ty, Option<TypeNs>)

pub fn lower_lifetime(&self, lifetime: &LifetimeRef) -> Lifetime