Struct hir_ty::TyLoweringContext

pub struct TyLoweringContext<'a> {
    pub db: &'a dyn HirDatabase,
    pub resolver: &'a Resolver,
    pub type_param_mode: ParamLoweringMode,
    pub impl_trait_mode: ImplTraitLoweringMode,
    /* private fields */
}

Fields
db: &'a dyn HirDatabaseresolver: &'a Resolvertype_param_mode: ParamLoweringMode

Note: Conceptually, it’s thinkable that we could be in a location where some type params should be represented as placeholders, and others should be converted to variables. I think in practice, this isn’t possible currently, so this should be fine for now.

impl_trait_mode: ImplTraitLoweringMode
Implementations

impl<'a> TyLoweringContext<'a>

pub fn new(db: &'a dyn HirDatabase, resolver: &'a Resolver) -> Self

pub fn with_debruijn<T>(
&self,
debruijn: DebruijnIndex,
f: impl FnOnce(&TyLoweringContext<'_>) -> T
) -> T

pub fn with_shifted_in<T>(
&self,
debruijn: DebruijnIndex,
f: impl FnOnce(&TyLoweringContext<'_>) -> T
) -> T

pub fn with_impl_trait_mode(
self,
impl_trait_mode: ImplTraitLoweringMode
) -> Self

pub fn with_type_param_mode(self, type_param_mode: ParamLoweringMode) -> Self

impl<'a> TyLoweringContext<'a>

pub fn lower_ty(&self, type_ref: &TypeRef) -> Ty

pub fn lower_ty_ext(&self, type_ref: &TypeRef) -> (Ty, Option<TypeNs>)